### 2023-12-16 #6155 @editorialbot
[PRE REVIEW]: XRTpy : A Hinode X-Ray Telescope Python Package
**Submitting author:** @https://github.com/joyvelasquez (<a href="http://orcid.org/0009-0005-4804-0946">Joy velasquez</a>)
**Repository:** https://github.com/HinodeXRT/xrtpy
**Branch with paper.md** (empty if default branch): main
**Version:** v0.4.0

Thanks for submitting your paper to JOSS @https://github.com/joyvelasquez. **Currently, there isn't a JOSS editor assigned** to your paper.
