### 2024-01-19 bd9291d
Added make_exposure_map.py which can be used to create exposure maps for use with temperature_from_filter_ratio
