### 2024-01-23 5580120
Cleared cells of Remove_lightleak notebook. Created new notebook illustrating use of composite images with temperature_from_filter_ratio. Switched from using requests to using astropy.utils.data.download_file
