+++ .pre-commit-config.yaml
@@ -52,7 +52,7 @@ repos:
   hooks:
   - id: nbqa-black
     additional_dependencies:
-    - black==23.3.0
+    - black==24.1.1
     args:
     - --nbqa-mutate
   - id: nbqa-isort
