+++ .pre-commit-config.yaml
@@ -8,14 +8,17 @@ repos:
   rev: v4.5.0
   hooks:
   - id: check-ast
+    name: validate Python code
+  - id: check-merge-conflict
+    name: check for git merge conflicts
+    exclude: .*\.rst
+  - id: check-case-conflict
+    name: check for filename case conflicts
+  - id: name-tests-test
+    args: [--pytest-test-first]
+    exclude: plasmapy/tests/_helpers/.*
   - id: trailing-whitespace
   - id: end-of-file-fixer
-    exclude: .*\.fits
-  - id: check-merge-conflict
-    exclude: .*\.rst
-  - id: requirements-txt-fixer
-  - id: check-symlinks
-  - id: check-case-conflict
   - id: check-json
   - id: check-toml
   - id: check-yaml
