+++ .pre-commit-config.yaml
@@ -50,7 +50,8 @@ repos:
     args: [--autofix]
     # For the labeler GitHub Action, labels with spaces in them must
     # be put in quotes. However, the pretty-format-yaml hook will
-    # remove the quotes which will break that action.
+    # remove the quotes which will break that action (and certain other
+    # actions).
     exclude: .github/labeler.yml|.pre-commit-search-and-replace.yaml
 
 - repo: https://github.com/MarcoGorelli/absolufy-imports
@@ -59,20 +60,67 @@ repos:
   - id: absolufy-imports
     name: Use absolute rather than relative imports
 
-
-
-
-
-- repo: https://github.com/tox-dev/pyproject-fmt
-  rev: 1.7.0
+- repo: https://github.com/pre-commit/pygrep-hooks
+  rev: v1.10.0
   hooks:
-  - id: pyproject-fmt
+  - id: python-check-blanket-noqa
+    name: noqa comments have an error code
+    exclude: ^docs/plasmapy_sphinx/
+  - id: rst-directive-colons
+  - id: rst-inline-touching-normal
+  - id: text-unicode-replacement-char
+
+
+- repo: https://github.com/codespell-project/codespell
+  rev: v2.2.6
+  hooks:
+  - id: codespell
+    name: codespell (add false positives to pyproject.toml)
+    args: [--write-changes]
+    additional_dependencies:
+    - tomli
+    exclude: .*\.fits
 
 - repo: https://github.com/astral-sh/ruff-pre-commit
   rev: v0.1.15
   hooks:
   - id: ruff
+    name: ruff (see https://docs.astral.sh/ruff/rules)
     args: [--fix]
+  - id: ruff-format
+    name: autoformat source code with ruff formatter
+
+- repo: https://github.com/asottile/blacken-docs
+  rev: 1.16.0
+  hooks:
+  - id: blacken-docs
+    name: autoformat code blocks in docs
+    exclude: docs/contributing/coding_guide.rst
+    additional_dependencies:
+    - black==24.1.1
+
+- repo: https://github.com/nbQA-dev/nbQA
+  rev: 1.7.1
+  hooks:
+  - id: nbqa-check-ast
+    name: validate Python notebooks
+  - id: nbqa-ruff
+    name: ruff for notebooks (see https://docs.astral.sh/ruff/rules)
+    args: [--fix, '--select=A,ARG,B,BLE,C,C4,E,F,FLY,I,INT,ISC,PERF,PIE,PLC,PLE,PYI,Q003,RET,RSE,SIM,TID,TRY,UP,W', '--ignore=B007,B008,B015,B018,E402,E501,E741,SIM114']
+
+- repo: https://github.com/astral-sh/ruff-pre-commit
+  rev: v0.1.15
+  hooks:
+  - id: ruff
+    name: ruff (see https://docs.astral.sh/ruff/rules)
+    args: [--fix]
+  - id: ruff-format
+    name: autoformat source code with ruff formatter
+
+
+
+
+
 
 - repo: https://github.com/PyCQA/isort
   rev: 5.13.2
@@ -85,46 +133,17 @@ repos:
     types:
     - python
 
-- repo: https://github.com/psf/black
-  rev: 24.1.1
-  hooks:
-  - id: black
-
 - repo: https://github.com/nbQA-dev/nbQA
   rev: 1.7.1
   hooks:
   - id: nbqa-black
     additional_dependencies:
     - black==24.1.1
-    args:
-    - --nbqa-mutate
   - id: nbqa-isort
     additional_dependencies:
     - isort==5.12.0
-    args:
-    - --nbqa-mutate
 
[…]
