+++ .pre-commit-config.yaml
@@ -122,17 +122,6 @@ repos:
 
 
 
-- repo: https://github.com/PyCQA/isort
-  rev: 5.13.2
-  hooks:
-  - id: isort
-    name: isort
-    entry: isort
-    require_serial: true
-    language: python
-    types:
-    - python
-
 - repo: https://github.com/nbQA-dev/nbQA
   rev: 1.7.1
   hooks:
