### 2024-02-01 9148f9e
Apply hooks to docs/conf.py
