+++ xrtpy/response/tests/test_temperature_response.py
@@ -1,7 +1,7 @@
-import pytest
-
 from datetime import datetime
 from pathlib import Path
+
+import pytest
 
 from xrtpy.response.temperature_response import TemperatureResponseFundamental
 
