### 2024-02-01 4b9a4db
Remove unused # noqa comment
