+++ xrtpy/__init__.py
@@ -8,7 +8,7 @@ import warnings
 from xrtpy import response
 
 try:
-    from xrtpy.version import __version__  # noqa
+    from xrtpy.version import __version__
 except ImportError:
     warnings.warn("version not found.")
 
