### 2024-02-01 219527f
Remove unused variable definitions in a test
