+++ xrtpy/response/tests/test_temperature_from_filter_ratio.py
@@ -134,9 +134,6 @@ def test_binning_case():
     goodT = (T_e.data > 0.0) & (idlTe > 0.0)
     goodE = (EM.data > 0.0) & (idlEM > 0.0)
 
-    delta = 10.0 ** T_e.data[goodT] - 10.0 ** idlTe[goodT]  # noqa
-    x = 10.0 ** idlTe[goodT]  # noqa
-
     assert np.allclose(
         10.0 ** T_e.data[goodT], 10.0 ** idlTe[goodT], atol=2.0e5, rtol=0.02
     )
