### 2024-02-01 71dc088
Ignore fewer ruff rules in notebooks
