+++ pyproject.toml
@@ -10,6 +10,13 @@ requires = [
 target-version = "py39"
 show-fixes = true
 extend-select = [
+  "ARG",
+  "B",
+  "BLE",
+  "C4",
+  "C90",
+  "COM818",
+  "FBT003",
   "I", # isort
   "UP", # pyupgrade
 ]
@@ -23,15 +30,18 @@ extend-exclude = [
   "_dev",
 ]
 
+[tool.ruff.per-file-ignores]
+"__init__.py" = ["E402", "F401", "F402", "F403"]  # ignore import errors
+"docs/notebooks/computing_functions/temperature_response.ipynb" = ["A001"]  # filter variable shadows Python builtin
+
 [tool.ruff.flake8-import-conventions.aliases]
 "astropy.units" = "u"
 "matplotlib.pyplot" = "plt"
 numpy = "np"
 pandas = "pd"
 
-[tool.ruff.per-file-ignores]
-"__init__.py" = ["E402", "F401", "F402", "F403"]  # ignore import errors
-"docs/notebooks/computing_functions/temperature_response.ipynb" = ["A001"]  # filter variable shadows Python builtin
+[tool.ruff.mccabe]
+max-complexity = 12
 
 [tool.codespell]
 skip = "*.genx,*.geny,*.png,*egg*,.git,.hypothesis,.nox,.tox,.idea,__pycache__,_build"
