### 2024-02-02 cd1ef5c
Update ruff rules
