+++ pyproject.toml
@@ -9,19 +9,64 @@ requires = [
 [tool.ruff]
 target-version = "py39"
 show-fixes = true
+show-source = true
+# Find info about ruff rules at: https://docs.astral.sh/ruff/rules
 extend-select = [
-  "ARG",
-  "B",
-  "BLE",
-  "C4",
-  "C90",
-  "COM818",
-  "FBT003",
+  "ARG", # flake8-unused-arguments
+  "B", # flake8-bugbear
+  "BLE", # flake8-blind-except
+  "C4", # flake8-comprehensions
+  "C90", # mccabe
+  "COM818", # trailing-comma-on-bare-tuple
+  "FBT003", # flake8-boolean-trap
+  "FLY", # flynt
   "I", # isort
+  "ICN", # flake8-import-conventions
+  "INP", # flake8-no-pep420
+  "INT", # flake8-gettext
+  "ISC", # flake8-implicit-str-concat
+  "N", # pep8-naming
+  "NPY", # numpy-deprecated-type-alias
+  "PD", # pandas-vet
+  "PERF", # perflint
+  "PGH", # pygrep-hooks
+  "PIE", # flake8-pie
+  "PLC", # pylint convention
+  "PLE", # pylint errors
+  "PLW", # pylint warnings
+  "PT", # flake8-pytest-style
+  "PTH", # flake8-use-pathlib
+  "PYI", # flake8-pyi
+  "RSE", # flake8-raise
+  "RUF005",# collection-literal-concatenation
+  "RUF006", # asyncio-dangling-task
+  "RUF007", # pairwise-over-zipped
+  "RUF008", # mutable-dataclass-default
+  "RUF009", # function-call-in-dataclass-default-argument
+  "RUF010", # explicit-f-string-type-conversion
+  "RUF011", # static-key-dict-comprehension
+  "RUF013", # implicit-optional
+  "RUF015", # unnecessary-iterable-allocation-for-first-element
+  "RUF016", # invalid-index-type
+  "RUF100", # unused-noqa
+  "RUF200", # invalid-pyproject-toml
+  "S", # flake8-bandit
+  "SIM", # flake8-simplify
+  "TCH", # flake8-type-checking
+  "TID", # flake8-tidy-imports
+  "TRY", # tryceratops
   "UP", # pyupgrade
+  "W", # pycodestyle warnings
 ]
 ignore = [
-  "E501",
+  "E501", # line-too-long
+  "N802", # invalid-function-name
+  "N803", # invalid-argument-name
+  "N806", # non-lowercase-variable-in-function
+  "N816", # mixed-case-variable-in-global-scope
+  "S101", # asserts
+  "SIM108", # if-else-block-instead-of-if-exp
+  "TRY003", # raise-vanilla-args
 ]
 extend-exclude = [
   ".jupyter",
@@ -29,6 +74,8 @@ extend-exclude = [
   "_build",
   "_dev",
 ]
+namespace-packages = [".github/workflows", "docs"]
+
 
 [tool.ruff.per-file-ignores]
 "__init__.py" = ["E402", "F401", "F402", "F403"]  # ignore import errors
