### 2024-02-02 8169e6f
Apply ruff fixes and add noqa comments
