### 2024-02-02 4d81b30
Minor updates to pre-commit config
