+++ .pre-commit-config.yaml
@@ -16,7 +16,6 @@ repos:
     name: check for filename case conflicts
   - id: name-tests-test
     args: [--pytest-test-first]
-    exclude: plasmapy/tests/_helpers/.*
   - id: trailing-whitespace
   - id: end-of-file-fixer
   - id: check-json
@@ -65,11 +64,9 @@ repos:
   hooks:
   - id: python-check-blanket-noqa
     name: noqa comments have an error code
-    exclude: ^docs/plasmapy_sphinx/
   - id: rst-directive-colons
   - id: rst-inline-touching-normal
   - id: text-unicode-replacement-char
-
 
 - repo: https://github.com/codespell-project/codespell
   rev: v2.2.6
@@ -95,7 +92,6 @@ repos:
   hooks:
   - id: blacken-docs
     name: autoformat code blocks in docs
-    exclude: docs/contributing/coding_guide.rst
     additional_dependencies:
     - black==24.1.1
 
@@ -107,10 +103,6 @@ repos:
   - id: nbqa-ruff
     name: ruff for notebooks (see https://docs.astral.sh/ruff/rules)
     args: [--fix, '--select=A,ARG,B,BLE,C,C4,E,F,FLY,I,INT,ISC,PERF,PIE,PLC,PLE,PYI,Q003,RET,RSE,SIM,TID,TRY,UP,W', '--ignore=B018,E402,E501']
-
-- repo: https://github.com/nbQA-dev/nbQA
-  rev: 1.7.1
-  hooks:
   - id: nbqa-black
     additional_dependencies:
     - black==24.1.1
