### 2024-02-02 da0a944
Update pre-commit configuration (#216)
* Update standard pre-commit hooks

* Update .pre-commit-config.yaml

* Use pre-commit hook to consistentify line endings

* Update pre-commit config

* Use isort from pre-commit

* Apply pre-commit hooks to notebooks

* Apply hooks to docs/conf.py

* Apply ruff's isort rule set instead of isort

* Ignore Untitled*.ipynb in .gitignore

* Update pre-commit hooks

* Change filter → filter_ to avoid conflict with Python built-in

* Update pyproject.toml

* Apply pre-commit hooks to notebooks

* Remove unused # noqa comment

* Remove unused variable definitio
[…]
