+++ docs/conf.py
@@ -5,8 +5,8 @@
 # http://www.sphinx-doc.org/en/master/config
 
 import os
-
 from datetime import datetime
+
 from sphinx.application import Sphinx
 
 # -- Project information -----------------------------------------------------
@@ -205,11 +205,11 @@ nitpick_ignore_regex = [
 # be used globally.
 rst_epilog = ""
 for epilog_file in ["_links.rst", "_substitutions.rst"]:
-    with open(epilog_file) as file:
+    with open(epilog_file) as file:  # noqa: PTH123
         rst_epilog += file.read()
 
 
 def setup(app: Sphinx) -> None:
-    app.add_config_value("revision", "", True)
+    app.add_config_value("revision", "", True)  # noqa: FBT003
     app.add_css_file("css/admonition_color_contrast.css")
     app.add_css_file("css/plasmapy.css", priority=600)
