### 2024-02-02 a6e4269
Update GitHub Action to install pandoc from repo
