### 2024-02-02 e87ac93
Make GitHub Action more concise
