### 2024-02-02 0e2ce80
Remove upper limit on nbconvert
