+++ requirements/docs.txt
@@ -8,7 +8,7 @@ docutils >= 0.18
 ipykernel
 ipython
 jinja2 != 3.1
-nbconvert < 7.14
+nbconvert
 nbsphinx >= 0.9
 numpydoc >= 1.2
 pillow
