+++ setup.cfg
@@ -64,7 +64,7 @@ docs =
     ipykernel
     ipython
     jinja2 != 3.1
-    nbconvert < 7.14
+    nbconvert
     nbsphinx >= 0.9
     numpydoc >= 1.2
     pillow
