### 2024-02-02 632ed78
Build docs on macOS rather than Ubuntu
