+++ .github/workflows/testing.yml
@@ -62,7 +62,7 @@ jobs:
 
   documentation:
     name: ${{ matrix.name }}
-    runs-on: ubuntu-latest
+    runs-on: macos-latest
 
     strategy:
       fail-fast: false
