+++ .github/workflows/testing.yml
@@ -89,15 +89,10 @@ jobs:
       run: python -m pip install --progress-bar off --upgrade nox
 
     - name: Install graphviz
-      run: sudo apt install graphviz
+      run: brew install graphviz
 
     - name: Install pandoc
-      run: |
-        PANDOC_VERSION="3.1.11.1"
-        wget -q https://github.com/jgm/pandoc/releases/download/${PANDOC_VERSION}/pandoc-${PANDOC_VERSION}-1-amd64.deb
-        sudo dpkg -i pandoc-${PANDOC_VERSION}-1-amd64.deb
-        rm pandoc-${PANDOC_VERSION}-1-amd64.deb
-        pandoc --version
+      run: brew install pandoc
 
     - name: Build documentation
       run: nox -e build_docs_nitpicky -- -q
