### 2024-02-02 c058ec6
Revert "Use brew to install graphviz & pandoc"
This reverts commit b72ac708eed75408a3625c78322e9f88c5105539.
