### 2024-02-02 514858d
Revert "Build docs on macOS rather than Ubuntu"
This reverts commit 632ed78f36fb1abc1a4dec74993aab2266620989.
