### 2024-02-02 b688c16
Revert "Remove upper limit on nbconvert"
This reverts commit 0e2ce80236eb746338b101595f50d99332af39de.
