+++ setup.cfg
@@ -64,7 +64,7 @@ docs =
     ipykernel
     ipython
     jinja2 != 3.1
-    nbconvert
+    nbconvert < 7.14
     nbsphinx >= 0.9
     numpydoc >= 1.2
     pillow
