+++ .pre-commit-config.yaml
@@ -79,7 +79,7 @@ repos:
     exclude: .*\.fits
 
 - repo: https://github.com/astral-sh/ruff-pre-commit
-  rev: v0.1.15
+  rev: v0.2.0
   hooks:
   - id: ruff
     name: ruff (see https://docs.astral.sh/ruff/rules)
