+++ .pre-commit-config.yaml
@@ -23,7 +23,7 @@ repos:
   - id: check-yaml
 
 - repo: https://github.com/python-jsonschema/check-jsonschema
-  rev: 0.27.4
+  rev: 0.28.0
   hooks:
   - id: check-github-workflows
 
@@ -79,7 +79,7 @@ repos:
     exclude: .*\.fits
 
 - repo: https://github.com/astral-sh/ruff-pre-commit
-  rev: v0.2.0
+  rev: v0.2.1
   hooks:
   - id: ruff
     name: ruff (see https://docs.astral.sh/ruff/rules)
