### 2024-02-08 e6ca974
[pre-commit.ci] pre-commit autoupdate (#219)
* [pre-commit.ci] pre-commit autoupdate

updates:
- [github.com/astral-sh/ruff-pre-commit: v0.1.15 → v0.2.0](https://github.com/astral-sh/ruff-pre-commit/compare/v0.1.15...v0.2.0)

* Another pre-commit autoupdate

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Nick Murphy <namurphy@cfa.harvard.edu>
