### 2024-02-08 0f7855d
Create check_contamination_data.yml
Create a GitHub Action to automatically update the data files,address issue #80
