### 2024-02-09 49751ef
Add Sphinx extensions (#187)
* Add Sphinx extensions

* Add docs requirements

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update requirements/docs.txt

* Add changelog entry

* Change _ to - in extension names

* Update requirements

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
