+++ setup.cfg
@@ -71,6 +71,7 @@ docs =
     pygments >= 2.11
     sphinx >= 6.0
     sphinx-changelog >= 1.3
+    sphinx-codeautolink >= 0.15.0
     sphinx-copybutton >= 0.5
     sphinx-gallery >= 0.12
     sphinx-hoverxref >= 1.1
