### 2024-02-10 #6155 @dfm
👋 @dhuppenkothen, @tddesjardins, @dpshelio — Would any of you be available and willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
