### 2024-02-12 5e06930
Adding to @Jslavin Pr 207 - Add optional use of exposure maps to temperature_from_filter_ratio #207 (#209)
* Fix paths to spectral model files in temperature response

The path to the model for hybrid abundances and photospheric abundances were switched.

* Fix atol in test_temperature_response_coronal.py

Made atol scaled to temperature response to provide a more stringent test.

* Fix atol in test_temperature_response_hybrid.py

Made atol scaled to temperature response to provide a more stringent test.

* Fix atol in test_temperature_response_photospheric.py

Made atol scaled to temperature 
[…]
