### 2024-02-13 #6155 @dfm
@dpshelio — Thanks for the response! It would be absolutely fine for you to not start your review for 2 weeks. I should note that JOSS reviews are iterative (reviewers give feedback and authors act on that in real time) so there might even be some small steps that you could start on with limited effort before that. But regardless, I'll be watching the discussion throughout the process so just keep us all posted on your capacity as we go and we'll make it work. Many thanks!!
