### 2024-02-13 6063eb8
Update ruff configuration
