### 2024-02-13 cb297e1
Update ruff configuration
