### 2024-02-13 048ff3d
Update pre-commit configuration (#223)
* Apply ruff changes

* Update ruff configuration

* Update ruff configuration
