### 2024-02-22 #6155 @dfm
:wave: @Cadair, @tyakobchuk — Would either of you be available and willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
