### 2024-02-22 #6396 @editorialbot
[REVIEW]: XRTpy : A Hinode X-Ray Telescope Python Package
**Submitting author:** @joyvelasquez (<a href="http://orcid.org/0009-0005-4804-0946">Joy velasquez</a>)
**Repository:** https://github.com/HinodeXRT/xrtpy
**Branch with paper.md** (empty if default branch): main
**Version:** 0.4.1
