+++ xrtpy/util/filter_name_processing.py
@@ -1,0 +1,1 @@
+# Creating the option for users to enter two XRT filter channel name as an input
