### 2024-02-23 660a539
Temperature notebook filter ratio method (#225)
* Python script for notebook

* notebook example
