+++ setup.cfg
@@ -45,6 +45,8 @@ extras =
     codespell
     pre-commit
     nox
+    imageio
+    ipywidgets
 tests =
     # ought to mirror requirements/tests.txt
     %(extras)s
