+++ setup.cfg
@@ -45,8 +45,6 @@ extras =
     codespell
     pre-commit
     nox
-    imageio
-    ipywidgets
 tests =
     # ought to mirror requirements/tests.txt
     %(extras)s
@@ -81,6 +79,8 @@ docs =
     sphinxcontrib-bibtex >= 2.5
     sunpy[net] >= 3.0.0
     towncrier >= 22.12
+    imageio
+    ipywidgets
 developer =
     # install everything for developers
     # ought to functionally mirror requirements.txt
