### 2024-03-27 6462973
Adding nbsphinx_allow_errors
