### 2024-03-28 HinodeXRT/xrtpy#227 @dpshelio
[JOSS-Review 6396] meta-issue
@joyvelasquez - I'm creating this issue to link all the issues and keep the review issue at JOSS without much noise. Last point is a set of suggestions and no requirements.
Review openjournals/joss-reviews/issues/6396

All good here.

- [x] #228

- [x] #229
- [x] #230
- [x] #231

- [x] #232
- [x] #233
- [x] Other improvements/suggestions
    - [x] UKSA didn't exist when Hinode was launched (by 3.5 years), though various institutions and a UK funding body were involved, I believed saying European includes the UK, as they were (and are) members of ESA.
      https://github.com/HinodeXRT/xrtpy/blob/660a539f3d6c7ee1e4596e86d4fdbbed5d74871e/joss/paper.md#L51
    - [x] for figures 1 and 2, there's a legend with a line and a date. I believe this indicates that this was the effective area and the response as per that date respectively. However, by having that in a legend and with a line, I wonder whether there's a black line that I'm missing. I recommend puting that as a graph title, or in a [legend without a line/symbol](https://stackoverflow.com/a/49507694/1087595).
    - [x] For figure 3 I would suggest to use a different colour table that's won't show the darker colours at each end. This is how that graph looks simulating achromatopsia. Google's [report about the goods of turbo color table](https://blog.research.google/2019/08/turbo-improved-rainbow-colormap-for.html) focuses on being good for depth or where there's a sequential. Here the red and blue happen too close to differenciate which one is which.
      ![sample image from documentation showing the simulated effect of achromatopsia, the colour bar is dark at the edges and clear in the middle, making impossible to know which "colour" refears to what value](https://github.com/HinodeXRT/xrtpy/assets/963242/fee46971-d3ad-428f-922c-98dbc797ed4c)
