### 2024-03-28 HinodeXRT/xrtpy#229 @dpshelio
[JOSS-Review 6396] Installation instructions
### Description

Should the developer installation be: `pip install -e ".[developer]"` (i.e., with `"`)? it fails for me otherwise.
