### 2024-03-28 HinodeXRT/xrtpy#230 @dpshelio
[JOSS-Review 6396] Functionality documentation
### Description

_Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?_

Some functions have very good documentation, e.g., [`xrtpy.response.temperature_from_filter_ratio.temperature_from_filter_ratio`](https://xrtpy.readthedocs.io/en/stable/api/xrtpy.response.temperature_from_filter_ratio.temperature_from_filter_ratio.html#xrtpy.response.temperature_from_filter_ratio.temperature_from_filter_ratio).

Some functions/methods/attributes have no documentation. For example: `xrtpy.response.TemperatureResponseFundamental().integration()`

The rest is somehow in between. It's quite difficult to get the way around the documentation as it's not provided in any order but as sphinx finds it.  I'd suggest providing some order through pages to make it easier for the contributors to find their way around it.
