### 2024-03-28 #6396 @dpshelio
> @dpshelio, @Cadair — This is a ping to please get started with your reviews ASAP. Many thanks!

Sorry for the delay. @dfm  I've set a number of issues on xrtpy as shown above. I will be away till 1st of May, so unfortunately I won't be able to review any comments till then. In any case, most of the comments are simple to address, and I trust the authors to fulfil them and by completing that meta issue I created will be OK with me to check the missing checkboxes above.
