+++ joss/paper.md
@@ -58,7 +58,7 @@ It offers a comprehensive range of functionalities, including object-oriented re
 These capabilities empower researchers to explore and analyze XRT data comprehensively, contributing to a deeper understanding of solar phenomena.
 
 The official analysis routines for Hinode are scripted in the Interactive Data Language (IDL).
-The SolarSoft XRT Analysis Guide serves as the official software and instrument guide for XRT data analysis.
+The [SolarSoft XRT Analysis Guide](https://xrt.cfa.harvard.edu/resources/documents/XAG/XAG.pdf) serves as the official software and instrument guide for XRT data analysis.
 XRTpy has been carefully written to ensure the consistency and replication of results obtained from the official IDL routines as described in the SolarSoft XRT Analysis Guide.
 Although currently XRTpy does not have all the capabilities of the SolarSoft routines for XRT, the package is in continual development and more functionality will be added in the future.
 This alignment with established practices and standards aims to facilitate a seamless transition for researchers while harnessing the benefits of Python in solar data analysis.
