### 2024-04-01 2d2c7c1
Updating/correcting Guidoni:2015 DOI
