+++ joss/paper.bib
@@ -78,7 +78,7 @@ Candle-flame-shaped Flare}",
   volume = {800},
   number = {54},
   pages = {},
-  doi = {10.1007/s11214-007-9173-7},
+  doi = { 10.1088/0004-637X/800/1/54},
   adsurl = {https://ui.adsabs.harvard.edu/abs/2015ApJ...800...54G/},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
