### 2024-04-02 fb995a0
Updated links for AIApy, EISpack, and IRISpy
