### 2024-04-02 2f2eb1d
Adding missing The SunPy Projec reference paper
