+++ joss/paper.bib
@@ -164,3 +164,18 @@ Data }",
   url = {https://www.aanda.org/articles/aa/abs/2015/01/aa24110-14/aa24110-14.html},
   note = {Published online: 11 December 2014}
 }
+
+@article{
+  title = {The SunPy Project: Open Source Development and Status of the Version 1.0 Core Package},
+  author = {SunPy Community; Barnes, Will T.; Bobra, Monica G.; Christe, Steven D.; Freij, Nabil; Hayes, Laura A.; Ireland, Jack; Mumford, Stuart; Perez-Suarez, David; Ryan, Daniel F.; Shih, Albert Y.; and others},
+  journal = {The Astrophysical Journal},
+  volume = {890},
+  pages = {68},
+  year = {2020},
+  doi = {10.3847/1538-4357/ab4f7a},
+  number = {1},
+  section = {},
+  publisher = {},
+  url = {https://ui.adsabs.harvard.edu/abs/2020ApJ...890...68S/abstract},
+  note = {}
+}
