### 2024-04-03 HinodeXRT/xrtpy#229 @Cadair
It depends on your shell if you need that extra escaping or not (it works in bash without but not zsh for example).
