### 2024-04-03 0c012d9
Add ipywidgets to docs requirements
