+++ requirements/docs.txt
@@ -7,6 +7,7 @@
 docutils >= 0.18
 ipykernel
 ipython
+ipywidgets
 jinja2 != 3.1
 nbconvert < 7.14
 nbsphinx >= 0.9
