### 2024-04-03 HinodeXRT/xrtpy#230 @Cadair
I agree with this that the API documentation could use some improvements. I don't even see how to navigate to that link @dpshelio provided from the nav bar of the docs. I would like to see an API section of the docs listing all the API. You may want to look into autoapi or automodapi to help with that.
