### 2024-04-03 #6396 @Cadair
I have opened a couple of issues and one PR on XRTPy, ans also agreed with some of @dpshelio's comments. I think all the things are pretty minor though. The package looks great and really useful. Would love to see it become SunPy affiliated as well.
