### 2024-04-09 #6396 @joyvelasquez
Hi @Cadair,

First off, thank you for taking on the role of reviewer! Additionally, I appreciate the issues you've created and the feedback provided, regardless of their scale. They'll undoubtedly contribute to our improvement in attention to detail. I'll be addressing them promptly. Also, a special thanks for recognizing the potential for XRTpy to become a Sunpy affiliated package!
