+++ joss/paper.bib
@@ -165,7 +165,7 @@ Data }",
   note = {Published online: 11 December 2014}
 }
 
-@article{
+@article{SunPy:2020,
   title = {The SunPy Project: Open Source Development and Status of the Version 1.0 Core Package},
   author = {SunPy Community; Barnes, Will T.; Bobra, Monica G.; Christe, Steven D.; Freij, Nabil; Hayes, Laura A.; Ireland, Jack; Mumford, Stuart; Perez-Suarez, David; Ryan, Daniel F.; Shih, Albert Y.; and others},
   journal = {The Astrophysical Journal},
@@ -177,5 +177,4 @@ Data }",
   section = {},
   publisher = {},
   url = {https://ui.adsabs.harvard.edu/abs/2020ApJ...890...68S/abstract},
-  note = {}
 }
