### 2024-05-23 c2fdb4a
JOSS Review Updates (#235)
* Adding SolarSoft XRT Analysis Guide reference link

* Updating/correcting Guidoni:2015 DOI

* Updating Python Package Index link

* Updated links for AIApy, EISpack, and IRISpy

* Adding missing The SunPy Projec reference paper

* Updating Sunpy reference

* referencing sunpy correctly
