+++ joss/paper.bib
@@ -78,7 +78,7 @@ Candle-flame-shaped Flare}",
   volume = {800},
   number = {54},
   pages = {},
-  doi = {10.1007/s11214-007-9173-7},
+  doi = { 10.1088/0004-637X/800/1/54},
   adsurl = {https://ui.adsabs.harvard.edu/abs/2015ApJ...800...54G/},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
@@ -163,4 +163,18 @@ Data }",
   publisher = {EDP Sciences},
   url = {https://www.aanda.org/articles/aa/abs/2015/01/aa24110-14/aa24110-14.html},
   note = {Published online: 11 December 2014}
+}
+
+@article{SunPy:2020,
+  title = {The SunPy Project: Open Source Development and Status of the Version 1.0 Core Package},
+  author = {SunPy Community; Barnes, Will T.; Bobra, Monica G.; Christe, Steven D.; Freij, Nabil; Hayes, Laura A.; Ireland, Jack; Mumford, Stuart; Perez-Suarez, David; Ryan, Daniel F.; Shih, Albert Y.; and others},
+  journal = {The Astrophysical Journal},
+  volume = {890},
+  pages = {68},
+  year = {2020},
+  doi = {10.3847/1538-4357/ab4f7a},
+  number = {1},
+  section = {},
+  publisher = {},
+  url = {https://ui.adsabs.harvard.edu/abs/2020ApJ...890...68S/abstract},
 }
