### 2024-05-23 971b5c3
Add the nbsphinx_allow_errors option
