### 2024-05-24 cafef00
Removing UK since it already mention Europe
