### 2024-05-24 0a0722a
Fixing Formatting problems in XRTpy JOSS md paper (#242)
* Properly rendering numbers

* Removing UK since it already mention Europe

* color map that is more suitable for more viewers

* Correcting misleading line in plot

* Correcting misleading line in plot

* Adding @ to @Fludra-and-Schmelz:1999

* Removing extra Figure words
