+++ joss/paper.bib
@@ -34,6 +34,7 @@ Yohkoh-BCS flare spectra}",
   doi = {10.1086/191698}
 }
 
+
 @ARTICLE{Golub:2007,
   author = {{Golub}, L. and {Deluca}, E. and {Austin}, G. and {Bookbinder}, J.
 and {Caldwell}, D. and {Cheimets}, P. and {Cirtain}, J. and {Cosmo}, M. and
