### 2024-05-24 68ffa73
Adding sunpy.map and requests as a install_requires
