### 2024-05-24 c28ca79
Removing sunpy.map
