+++ setup.cfg
@@ -39,7 +39,6 @@ install_requires =
     sunpy >= 3.0.0
     scikit-image >= 0.16.2
     requests
-    sunpy.map
 
 [options.extras_require]
 extras =
