### 2024-05-24 c743afc
Adding sunpy map
