+++ setup.cfg
@@ -38,6 +38,7 @@ install_requires =
     scipy >= 1.7.0
     sunpy >= 3.0.0
     scikit-image >= 0.16.2
+    sunpy[map] >= 3.0.0
     requests
 
 [options.extras_require]
