### 2024-05-24 99d0f57
Addressing JOSS installation issues (#247)
* Adding sunpy.map and requests as a install_requires

* Removing sunpy.map

* Adding sunpy map

* Update Documentation
