+++ docs/contributing/contributing.rst
@@ -1,0 +1,7 @@
+.. _contributing:
+
+*********************
+Contributing to XRTpy
+*********************
+
+Thank you for your interest in contributing to XRTpy! We welcome contributions from the community to improve and expand the functionality of this package.
