### 2024-05-24 dd6c8a6
Removing workflow.rst and adding contributing
