### 2024-05-24 9ae0e64
Adding section for  Reporting Issues
