+++ docs/contributing/contributing.rst
@@ -5,3 +5,7 @@ Contributing to XRTpy
 *********************
 
 Thank you for your interest in contributing to XRTpy! We welcome contributions from the community to improve and expand the functionality of this package.
+
+There are several ways you can contribute to XRTpy:
+
+1. **Reporting Issues**: If you encounter any bugs or have suggestions for improvements, please report them using the GitHub [Issues](https://github.com/HinodeXRT/xrtpy/issues) page. Provide as much detail as possible, including steps to reproduce the issue and any relevant screenshots or code snippets.
