+++ docs/contributing/contributing.rst
@@ -9,3 +9,11 @@ Thank you for your interest in contributing to XRTpy! We welcome contributions f
 There are several ways you can contribute to XRTpy:
 
 1. **Reporting Issues**: If you encounter any bugs or have suggestions for improvements, please report them using the GitHub [Issues](https://github.com/HinodeXRT/xrtpy/issues) page. Provide as much detail as possible, including steps to reproduce the issue and any relevant screenshots or code snippets.
+
+2. **Submitting Pull Requests**: If you want to contribute code, follow these steps:
+   - Fork the repository on GitHub.
+   - Create a new branch from the `main` branch for your changes.
+   - Make your changes, ensuring that you follow the coding standards and include tests for any new functionality.
+   - Commit your changes with clear and descriptive commit messages.
+   - Push your branch to your forked repository.
+   - Create a pull request (PR) from your branch to the `main` branch of the original repository.
