+++ docs/contributing/contributing.rst
@@ -20,10 +20,11 @@ There are several ways you can contribute to XRTpy:
 
 3. **Improving Documentation**: Good documentation is crucial for the usability of the package. You can help by improving existing documentation, or writing new tutorials.
 
-## Development Environment
 
+Development Environment
 To set up your development environment:
 
 1. Clone the repository:
-   ```git clone https://github.com/HinodeXRT/xrtpy.git```
-   ```cd xrtpy```
+   ```bash
+   git clone https://github.com/HinodeXRT/xrtpy.git
+   cd xrtpy
