### 2024-05-24 af931f9
Fixing method of text
