### 2024-05-24 d22078f
Adding list of methods to contribute
