+++ docs/contributing/contributing.rst
@@ -28,20 +28,20 @@ To set up your development environment:
 
 1. Clone the repository::
 
-   `git clone https://github.com/HinodeXRT/xrtpy.git`
-   `cd xrtpy`
+   git clone https://github.com/HinodeXRT/xrtpy.git
+   cd xrtpy
 
 2. Install the required dependencies::
 
-   `pip install -r requirements.txt`
+   pip install -r requirements.txt
 
 3. Install the package in editable mode::
 
-   `pip install -e .`
+   pip install -e .
 
 Coding Standards
 ================
-- Follow the PEP 8 coding style.
+- Follow the 'PEP-8-coding-style'_.
 - Write clear and concise commit messages.
 - Include docstrings for all functions and classes.
 - Ensure that your code is covered by tests and that all tests pass before submitting a PR.
@@ -51,7 +51,7 @@ Testing
 =======
 We use `pytest` for testing. To run the tests, use the following command::
 
-   `pytest`
+   pytest
 
 Ensure that all tests pass before submitting your PR.
 
