+++ docs/contributing/index.rst
@@ -6,6 +6,7 @@ Contributor Guide
 
 .. toctree::
    :maxdepth: 3
+
    contributing
    installation_for_development
    release_guide
