### 2024-05-25 6813dfe
Updating contributing section (#248)
* Adding a contributing section

* Removing workflow.rst and adding contributing

* Adding section for  Reporting Issues

* Adding section for Submitting Pull Requests

* Adding a section for Improving Documentation

* Adding method to set up your development environment

* Fixing format

* Fixing method of text

* Updating text

* Adding list of methods to contribute

* Updating format to pass git doc test

* Updating format

* Updating format

* Finally updating

* Removed Development Environment

* Removed IDL papge, and changed format

* Updating name


[…]
