+++ docs/contributing/install_dev.rst
@@ -1,7 +1,0 @@
-.. _installation for development:
-
-****************************
-Installation for Development
-****************************
-
-This page has not yet been written.
