+++ docs/contributing/contributing.rst
@@ -1,0 +1,25 @@
+.. _contributing:
+
+*********************
+Contributing to XRTpy
+*********************
+
+Thank you for your interest in contributing to XRTpy! We welcome contributions from the community to improve and expand the functionality of this package.
+
+There are several ways you can contribute to XRTpy:
+
+1. **Reporting Issues**: If you encounter any bugs or have suggestions for improvements, please report them using the `GitHub-issue`_ page. Provide as much detail as possible, including steps to reproduce the issue and any relevant screenshots or code snippets.
+
+2. **Submitting Pull Requests**: If you want to contribute code, follow these steps:
+   - Fork the repository on GitHub.
+   - Create a new branch from the `main` branch for your changes.
+   - Make your changes, ensuring that you follow the coding standards and include tests for any new functionality.
+   - Commit your changes with clear and descriptive commit messages.
+   - Push your branch to your forked repository.
+   - Create a pull request (PR) from your branch to the `main` branch of the original repository.
+
+3. **Improving Documentation**: Good documentation is crucial for the usability of the package. You can help by improving existing documentation, or writing new tutorials.
+
+Visit the `Development Environment` section to set up your development environment. Thank you for contributing to XRTpy!
+
+.. _GitHub-issue: https://github.com/HinodeXRT/xrtpy/issues
