### 2024-05-25 #6396 @joyvelasquez
@dfm, sorry for the late response. I'm almost done updating all the suggestions. I just have two more issues to address: [XRTpy Issue #227](https://github.com/HinodeXRT/xrtpy/issues/227) and [Issue #229](https://github.com/HinodeXRT/xrtpy/issues/229). I plan to finish them early next week. I'll make sure to let you know!
