### 2024-05-28 bfd4743
Updating get_abundance_data with underscore for internal use
