### 2024-05-28 1f05a8b
Updating/add docstring infromation
