### 2024-05-28 13b5ff1
Updating Quantity,quantity, and np.array
