### 2024-05-28 aefb5b3
Updating string,  np.array
