### 2024-05-28 c0c52f4
Updating quantity str
