### 2024-05-28 fa48386
Updated two functions
