### 2024-05-28 3fa32aa
Leaving contamination_on_CCD since intended for external use
