### 2024-05-28 68ab1ea
Adding docstring to effective_area function
