### 2024-05-28 d3c41c2
updated some doc strings and covert function to be used internally
