### 2024-05-28 d4819e3
Coverting function to be used internally
