### 2024-05-28 014d1ef
Coverting function to be used internally
