### 2024-05-28 0864682
Coverting function to be used internally-_filterwheel_angular_wavenumber
