### 2024-05-28 8c8242d
Coverting function to be used internally- _angular_wavenumber_CCD
