### 2024-05-28 5cc43ad
Coverting function to be used internally- _CCD_contamination_transmission
