### 2024-05-28 4d42952
Coverting function to be used internally- _combo_filter_name_split
