### 2024-05-28 99d6247
Coverting function to be used internally- _combo_filter1/2_wheel_number
