+++ xrtpy/response/effective_area.py
@@ -154,13 +154,13 @@ class EffectiveAreaFundamental:
         return filter1, filter2
 
     @property
-    def combo_filter1_wheel_number(self):
+    def _combo_filter1_wheel_number(self):
         """Defining chosen filter to its corresponding filter wheel."""
         filter1, _ = self._combo_filter_name_split
         return 0 if filter1 in index_mapping_to_fw1_name else 1
 
     @property
-    def combo_filter2_wheel_number(self):
+    def _combo_filter2_wheel_number(self):
         """Defining chosen filter to its corresponding filter wheel."""
         _, filter2 = self._combo_filter_name_split
         return 0 if filter2 in index_mapping_to_fw1_name else 1
@@ -189,14 +189,14 @@ class EffectiveAreaFundamental:
     def combo_filter1_data(self):
         """Collecting filter data."""
         return _filter_contamination[self.combo_filter_index_mapping_to_name_filter1][
-            self.combo_filter1_wheel_number
+            self._combo_filter1_wheel_number
         ]
 
     @property
     def combo_filter2_data(self):
         """Collecting filter data."""
         return _filter_contamination[self.combo_filter_index_mapping_to_name_filter2][
-            self.combo_filter2_wheel_number
+            self._combo_filter2_wheel_number
         ]
 
     @property
