### 2024-05-28 f2af2d3
Adding docstring infromation contamination_on_filter1/2_combo
