### 2024-05-28 f01f2b3
Updating docstring infromation
