### 2024-05-28 d1ee894
Added the :noindex: directive to the docstrings of the name and observation_date
