### 2024-05-28 afbb12f
Adding docstring infromation _combo_filter_index_mapping_to_name_filter1/2
