### 2024-05-28 d36ee54
Adding docstring infromation _combo_filter_index_mapping_to_name_filter1/2
