### 2024-05-28 0dba7ba
Coverting function to be used internally- _contamination_on_filter1/2_combo
