### 2024-05-29 cdb7ed0
Updating resolve_filter_name docstring
