### 2024-05-29 22927a8
Adding docstring for Geometry
