### 2024-05-29 664bfd1
Updating docstring for channel
