### 2024-05-29 bc037ac
Updating EntranceFilter docsting
