### 2024-05-29 bec6a2d
Updating Mirror docstring
