### 2024-05-29 08cc5cd
Updating Filter docstring
