### 2024-05-29 cb595cc
Updating CCD docstring
