### 2024-05-29 6aa4410
Adding :noindex: to geometry docstring functions
