### 2024-05-29 2b0b927
Adding :noindex: to EntranceFilter function docstrings
