### 2024-05-29 67ae9d8
Adding :noindex: to Filter functions docstrings
