### 2024-05-29 ffde1eb
Adding :noindex: to CCD functions docstrings
