+++ xrtpy/response/channel.py
@@ -94,32 +94,22 @@ class Geometry:
         self._geom_data = self._genx_file[self._channel_index]["GEOM"]
 
     @property
-    def geometry_name(self) -> str:
-        """
-        Hinode/XRT flight model geometry.
-
-        :noindex:
-        """
+    def geometry_name(
+        self,
+    ) -> str:
+        """Hinode/XRT flight model geometry."""
         return self._geom_data["LONG_NAME"]
 
     @property
     @u.quantity_input
     def geometry_focal_len(self) -> u.cm:
-        """
-        XRT flight model geometry focal length.
-
-        :noindex:
-        """
+        """XRT flight model geometry focal length."""
         return u.Quantity(self._geom_data["FOC_LEN"], u.cm)
 
     @property
     @u.quantity_input
     def geometry_aperture_area(self) -> u.cm**2:
-        """
-        XRT flight model geometry aperture area.
-
-        :noindex:
-        """
+        """XRT flight model geometry aperture area."""
         return u.Quantity(self._geom_data["APERTURE_AREA"], u.cm**2)
 
 
@@ -164,66 +154,38 @@ class EntranceFilter:
 
     @property
     def entrancefilter_density(self) -> u.g * u.cm**-3:
-        """
-        XRT entrance filter material density in g/cm³.
-
-        :noindex:
-        """
+        r"""XRT entrance filter material density in g/cm\ :sup:`3`\ ."""
         return u.Quantity(self._en_filter_data["DENS"], u.g * u.cm**-3)
 
     @property
     def entrancefilter_material(self) -> str:
-        """
-        XRT entrance filter material.
-
-        :noindex:
-        """
+        """XRT entrance filter material."""
         return self._en_filter_data["MATERIAL"]
 
     @property
     def entrancefilter_mesh_transmission(self):
-        """
-        Transmission of mesh filter substrate.
-
-        :noindex:
-        """
+        """Transmission of mesh filter substrate."""
         return self._en_filter_data["MESH_TRANS"]
 
     @property
     def entrancefilter_name(self) -> str:
-        """
-        Entrance filter name.
-
-        :noindex:
-        """
+        """Entrance filter name."""
         return self._en_filter_data["LONG_NAME"]
 
     @property
     def number_of_wavelengths(self):
-        """
-        Data number length.
-
-        :noindex:
-        """
+        """Data number length."""
         return self._en_filter_data["LENGTH"]
 
     @property
     def entrancefilter_substrate(self) -> str:
-        """
-        XRT entrance filter substrate.
-
-        :noindex:
-        """
+        """XRT entrance filter substrate."""
         return self._en_filter_data["SUBSTRATE"]
 
     @property
     @u.quantity_input
     def entrancefilter_wavelength(self) -> u.angstrom:
-        """
-        Array of wavelengths for entrance filter transmission in angstroms.
-
-        :noindex:
-        """
+        """Array of wavelengths for entrance filter transmission in angstroms."""
         return u.Quantity(self._en_filter_data["WAVE"], u.angstrom)[
             : self.number_of_wavelengths
         ]
@@ -231,20 +193,12 @@ class EntranceFilter:
     @property
     @u.quantity_input
     def entrancefilter_thickness(self) -> u.angstrom:
-        """
-        XRT entrance filter material thickness in angstroms.
-
-        :noindex:
-        """
+        """XRT entrance filter material thickness in angstroms."""
         return u.Quantity(self._en_filter_data["THICK"], u.angstrom)
 
     @property
     def entrancefilter_transmission(self):
-        """
-        Entrance filter transmission.
-
-        :noindex:
-        """
+        """Entrance filter transmission."""
[…]
