+++ xrtpy/response/channel.py
@@ -115,35 +115,12 @@ class Geometry:
 
 class EntranceFilter:
     """
-    Represents the entrance filter of the X-Ray Telescope (XRT) on the Hinode spacecraft.
+    Entrance filter properties.
 
-    The entrance filter covers the annular entrance aperture of the XRT, serving to reduce
-    both visible light and heat load entering the instrument. These filters are critical
-    for maintaining the necessary optical conditions inside the telescope.
+    Thin prefilters cover the narrow annular entrance aperture of the XRT serving two main purposes:
 
-    Parameters
-    ----------
-    index : int
-        The index of the channel within the GENX file that contains data specific to this filter.
-
-    Attributes
-    ----------
-    entrancefilter_density : astropy.units.Quantity
-        The density of the entrance filter material in g/cm³.
-    entrancefilter_material : str
-        The material composition of the entrance filter.
-    entrancefilter_mesh_transmission : float
-        The percentage transmission of the mesh part of the filter.
-    entrancefilter_name : str
-        The descriptive name of the entrance filter.
-    entrancefilter_substrate : str
-        The substrate material of the entrance filter.
-    entrancefilter_thickness : astropy.units.Quantity
-        The thickness of the entrance filter material measured in Angstroms.
-    entrancefilter_transmission : numpy.ndarray
-        The transmission efficiency of the entrance filter across different wavelengths.
-    entrancefilter_wavelength : astropy.units.Quantity
-        The wavelengths at which the transmission data of the filter are measured, in Angstroms.
+    1. Reduce the visible light entering the instrument.
+    2. Reduce the heat load in the instrument.
     """
 
     _genx_file = _genx_file
@@ -204,32 +181,10 @@ class EntranceFilter:
 
 class Mirror:
     """
-    Defines a grazing incidence mirror used in the X-Ray Telescope (XRT) for imaging in soft X-rays.
+    Grazing incidence mirror properties.
 
-    XRT utilizes a two-bounce telescope design requiring at least two mirrors. This class provides
-    properties of one of these mirrors.
-
-    Parameters
-    ----------
-    index : int
-        Index of the channel within the GENX data file.
-    mirror_number : int
-        Specifies whether this is the first or second mirror (1 or 2).
-
-    Attributes
-    ----------
-    mirror_density : astropy.units.Quantity
-        The mass density of the mirror in g/cm³.
-    mirror_graze_angle : astropy.units.Quantity
-        The graze angle of the mirror during operation, measured in degrees.
-    mirror_material : str
-        The material composition of the mirror.
-    mirror_name : str
-        The name or identifier of the mirror.
-    mirror_reflection : numpy.ndarray
-        The reflectance values of the mirror at different wavelengths.
-    mirror_wavelength : astropy.units.Quantity
-        The wavelengths at which the mirror's reflectance is measured, in Angstroms.
+    Grazing-incidence optics used for soft X-ray imaging generally require a minimum of two surfaces.
+    Since XRT is a two-bounce telescope, there are two mirror reflectivities.
     """
 
     _genx_file = _genx_file
@@ -286,34 +241,9 @@ class Mirror:
 
 class Filter:
     """
-    Represents one of the focal plane filters of the X-Ray Telescope (XRT) on the Hinode spacecraft,
-    which are mounted on two filter wheels.
+    X-ray filters using two filter wheels.
 
-    Parameters
-    ----------
-    index : int
-        Index of the channel within the GENX data file.
-    filter_number : int
-        Specifies the filter wheel position (1 or 2).
-
-    Attributes
-    ----------
-    filter_density : astropy.units.Quantity
-        The density of the filter material in g/cm³.
-    filter_material : str
[…]
