### 2024-05-29 ed3afa5
Go back some steps to find error moment
