+++ xrtpy/response/channel.py
@@ -71,20 +71,6 @@ def resolve_filter_name(name):
 class Geometry:
     """
     The physical geometric parameters of the X-Ray Telescope (XRT) on board the Hinode spacecraft.
-
-    Parameters
-    ----------
-    index : int
-        The index of the channel in the GENX file.
-
-    Attributes
-    ----------
-    geometry_name : str
-        Hinode/XRT flight model geometry.
-    geometry_focal_len : astropy.units.Quantity
-        XRT flight model geometry focal length in cm.
-    geometry_aperture_area : astropy.units.Quantity
-        XRT flight model geometry aperture area in cm^2.
     """
 
     _genx_file = _genx_file
