### 2024-05-29 b79c036
Removed channel Attributes in docstring to find error
