### 2024-05-29 de68082
Adding a single  :noindex: to the single attribute in the channel doc string
