### 2024-05-29 c435361
Adding a single  :noindex: to the single attribute in the channel doc string
