### 2024-05-29 a93a7bc
Adding a single  :noindex: to the single attribute in the channel doc string-test6
