### 2024-05-29 7a091d3
Adding a single  :noindex: to the single attribute in the channel doc string-test6
