+++ xrtpy/response/channel.py
@@ -405,6 +405,8 @@ class Channel:
         """
         Geometric parameters of the XRT channel.
 
+        .. note:: This property is not indexed to avoid duplication.
+
         :noindex:
         """
         return self._geometry
