### 2024-05-30 d88d787
Adding a single  :noindex: to the single attribute in the channel doc string-test7
