### 2024-05-30 2d3fb0b
Testing EntranceFilter docstring
