### 2024-05-30 249ff98
Updating channal docstring
