### 2024-05-30 216e9dc
Testing a doc string in the channel class, name
