### 2024-05-30 9afbbcf
Testing a doc string in the channel class, name- adding noindex
