### 2024-05-30 e70f03c
Testing a doc string in the channel class, name - test 3
