+++ xrtpy/response/channel.py
@@ -383,7 +383,7 @@ class Channel:
         Properties of the second filter.
     CCD
         Properties of the CCD.
-    name : str
+    name
         Name of XRT X-Ray channel.
     """
 
@@ -475,8 +475,6 @@ class Channel:
     def name(self) -> str:
         """
         Name of XRT X-Ray channel.
-
-        :noindex:
         """
         return self._channel_data["NAME"]
 
