### 2024-05-30 9161908
Testing a doc string in the channel class, name - test 4
