### 2024-05-30 711a930
Removing noindex in channel class dogstrings
