### 2024-05-30 fe85d88
Adding the rest of the channel functions to the docstring
