### 2024-05-30 8caf554
Adding Geometry docstring
