### 2024-05-30 cdee6af
Adding EntranceFilter docstring
