### 2024-05-30 c8b3725
Testing new method for doc string-Channel.EntranceFilter.entrancefilter_density
