### 2024-05-30 c795552
Fixed format to the EntranceFilter docsting
