### 2024-05-30 3aea2bc
Adding Mirror Attributes docstring
