### 2024-05-30 f5783e6
Adding Filter Attributes docstring
