### 2024-05-30 7e258ef
Adding CCD Attributes docstring
