### 2024-05-30 8f9a682
Updating method of doc string in Channel class
