### 2024-05-30 bd9eef7
Removing function to test some docstrings
