### 2024-05-30 807f58f
Trying a another method for the docstring in Channel
