+++ xrtpy/response/channel.py
@@ -597,51 +597,27 @@ class Channel:
 
     @property
     def name(self) -> str:
-        """
-        Name of XRT X-Ray channel.
-        :no-index:
-        """
         return self._channel_data["NAME"]
 
     @property
     @u.quantity_input
     def wavelength(self) -> u.angstrom:
-        """
-        Array of wavelengths for every X-ray channel in angstroms.
-        :no-index:
-        """
         return u.Quantity(self._channel_data["WAVE"], u.angstrom)[
             : self.number_of_wavelengths
         ]
 
     @property
     def transmission(self):
-        """
-        Channel transmission.
-        :no-index:
-        """
         return self._channel_data["TRANS"][: self.number_of_wavelengths]
 
     @property
     def number_of_wavelengths(self):
-        """
-        Data number length.
-        :no-index:
-        """
         return self._channel_data["LENGTH"]
 
     @property
     def observatory(self) -> str:
-        """
-        The spacecraft name - Hinode.
-        :no-index:
-        """
         return self._channel_data["OBSERVATORY"]
 
     @property
     def instrument(self) -> str:
-        """
-        Instrument - X-Ray Telescope -XRT.
-        :no-index:
-        """
         return self._channel_data["INSTRUMENT"]
