### 2024-05-30 c251492
Trying a another method for the docstring in Channel-test 2
