### 2024-05-31 207ba9f
Update api_static pages
