+++ docs/api_static/xrtpy.util.make_exposure_map.rst
@@ -1,0 +1,8 @@
+:orphan:
+
+`xrtpy.util.make_exposure_map`
+==============================
+
+.. currentmodule:: xrtpy.util.make_exposure_map
+
+.. automodapi:: xrtpy.response.make_exposure_map
