+++ xrtpy/image_correction/__init__.py
@@ -5,6 +5,8 @@ __all__ = [
     "remove_lightleak",
 ]
 
+from xrtpy.image_correction import deconvolve, remove_lightleak
+
 _SSW_MIRRORS = [
     "https://sohoftp.nascom.nasa.gov/solarsoft/",
     "https://hesperia.gsfc.nasa.gov/ssw/",
