+++ noxfile.py
@@ -26,11 +26,9 @@ def tests(session):
     session.run("pytest", *pytest_options)
 
 
-@nox.session
 def linters(session):
-    session.install("-r", "requirements/tests.txt")
-    flake8_options = ["--count", "--show-source", "--statistics"]
-    session.run("flake8", "xrtpy", *flake8_options, *session.posargs)
+    session.install("pre-commit")
+    session.run("pre-commit", "run", "--all-files", *session.posargs)
 
 
 @nox.session
