+++ setup.cfg
@@ -50,11 +50,6 @@ extras =
 tests =
     # ought to mirror requirements/tests.txt
     %(extras)s
-    dlint
-    flake8
-    flake8-absolute-import
-    flake8-rst-docstrings
-    flake8-use-fstring
     pydocstyle
     pytest >= 5.4.0
     pytest-allclose
@@ -106,109 +101,6 @@ doctest_optionflags =
 addopts = --doctest-modules --doctest-continue-on-failure --ignore=docs/conf.py
 filterwarnings =
     ignore:.*Creating a LegacyVersion.*:DeprecationWarning
-
-[flake8]
-convention = numpy
-
-extend-ignore =
-    D105,
-    D107,
-    D202,
-    D205,
-    D302,
-    D400,
-    D403,
-    E203,
-    E501,
-    E731,
-    F401,
-    F405,
-    W605,
-    RST210,
-    RST213,
-    RST305,
-    RST306
-
-exclude =
-    extern,
-    sphinx,
-    *test*,
-    *parsetab.py,
-    conftest.py,
-    docs/conf.py,
-    setup.py,
-    .jupyter
-
-# Use rst-roles and rst-directives to list roles and directives from
-# Sphinx and its extensions so that they don't get flagged when using
-# flake8-rst-docstrings.
-rst-roles =
-    abbr
-    any
-    attr
-    cite
-    cite:ct
-    cite:cts
-    cite:p
-    cite:ps
-    cite:t
-    cite:ts
-    class
-    command
-    confval
-    data
-    dfn
-    doc
-    download
-    envvar
-    eq
-    event
-    exc
-    file
-    func
-    guilabel
-    kbd
-    keyword
-    makevar
-    manpage
-    menuselection
-    meth
-    mod
-    numref
-    option
-    pep
-    program
-    ref
-    regexp
-    rst:dir
-    samp
-    term
-    token
-
-rst-directives =
-    codeauthor
-    confval
-    deprecated
-    event
-    highlight
-    hlist
-    index
-    literalinclude
-    nbgallery
-    only
-    rst:directive
-    sectionauthor
-    seealso
-    tabularcolumns
-    todo
-    versionadded
-    versionchanged
-
-enable-extensions =
-    # Look for strings that have {} in them but aren't f-strings.
-    # If there is a false positive from this in a file, put that in
-    # per-file-ignores.
-    FS003
 
 [coverage:run]
 
