### 2024-05-31 2ab203c
Move _SSW_MIRRORS out of __init__.py
