+++ xrtpy/image_correction/_ssw_mirrors.py
@@ -1,0 +1,6 @@
+"""Websites that are mirrors of SolarSoft IDL."""
+
+_SSW_MIRRORS = [
+    "https://sohoftp.nascom.nasa.gov/solarsoft/",
+    "https://hesperia.gsfc.nasa.gov/ssw/",
+]
