### 2024-05-31 0ac8597
Revert "Move _SSW_MIRRORS out of __init__.py"
This reverts commit 2ab203cc3e63eb50c126f5ea76387a1f308c665e.
