+++ xrtpy/image_correction/__init__.py
@@ -4,3 +4,8 @@ __all__ = [
     "deconvolve",
     "remove_lightleak",
 ]
+
+_SSW_MIRRORS = [
+    "https://sohoftp.nascom.nasa.gov/solarsoft/",
+    "https://hesperia.gsfc.nasa.gov/ssw/",
+]
