+++ xrtpy/image_correction/_ssw_mirrors.py
@@ -1,0 +1,8 @@
+"""Mirrors of SolarSoft IDL."""
+
+__all__ = []
+
+_SSW_MIRRORS = [
+    "https://sohoftp.nascom.nasa.gov/solarsoft/",
+    "https://hesperia.gsfc.nasa.gov/ssw/",
+]
