+++ xrtpy/image_correction/__init__.py
@@ -6,8 +6,3 @@ __all__ = [
 ]
 
 from xrtpy.image_correction import deconvolve, remove_lightleak
-
-_SSW_MIRRORS = [
-    "https://sohoftp.nascom.nasa.gov/solarsoft/",
-    "https://hesperia.gsfc.nasa.gov/ssw/",
-]
