### 2024-05-31 221933a
Use SSW_MIRROS from xrtpy.util
