+++ xrtpy/util/__init__.py
@@ -1,12 +1,15 @@
-from xrtpy.util import time
+from xrtpy.util import time, filename2repo_path, make_exposure_map
 from xrtpy.util.time import epoch
 
 __all__ = [
     "epoch",
     "time",
+    "filename2repo_path",
+    "make_exposure_map",
+    "SSW_MIRRORS",
 ]
 
-_SSW_MIRRORS = [
+SSW_MIRRORS = (
     "https://sohoftp.nascom.nasa.gov/solarsoft/",
     "https://hesperia.gsfc.nasa.gov/ssw/",
-]
+)
