### 2024-05-31 73b4a3d
Remove flake8 (#255)
* Remove flake8

* Decorate with @nox.session

* Move _SSW_MIRRORS out of __init__.py

* Revert "Move _SSW_MIRRORS out of __init__.py"

This reverts commit 2ab203cc3e63eb50c126f5ea76387a1f308c665e.

* Remove linters from CI
