+++ noxfile.py
@@ -28,9 +28,8 @@ def tests(session):
 
 @nox.session
 def linters(session):
-    session.install("-r", "requirements/tests.txt")
-    flake8_options = ["--count", "--show-source", "--statistics"]
-    session.run("flake8", "xrtpy", *flake8_options, *session.posargs)
+    session.install("pre-commit")
+    session.run("pre-commit", "run", "--all-files", *session.posargs)
 
 
 @nox.session
