+++ xrtpy/util/make_exposure_map.py
@@ -9,6 +9,7 @@ from xrtpy.util.filename2repo_path import filename2repo_path
 
 __all__ = ["make_exposure_map"]
 
+
 def make_exposure_map(comp_image_file, qualfiles=None, retsatpix=False, verbose=False):
     """
     Make an exposure map for a composite image. That is, using data from the
