### 2024-06-03 #6396 @joyvelasquez
Hi @dfm,
I have addressed [most](https://github.com/HinodeXRT/xrtpy/issues/227) of the reviewers' suggestions/issues and merged them into the XRTpy Python package. However, I still need to cover one more topic: issue [#227.](https://github.com/HinodeXRT/xrtpy/issues/230) The XRTpy team is working together to update this topic. It's difficult to judge how long it will take, but we're actively working on it. I will update you once complete.
