### 2024-06-03 6786b07
Adding hinode png as thumb nail to a notebook: nbsphinx_thumbnails
