### 2024-06-03 e073ab4
Notebook Guide to Data Extraction (#210)
* Notebook Guide to Data Extraction

* Update text - added year notebook made

* Adding a new section  Filtering FITS Files Through Selected Filter

* Adding a new section:Visualizing Solar Dynamics: Creating a Movie from XRT FITS Data

* Adding a new section:Filtering FITS Files Through Pixel Size

* Adding a new section:Navigating FITS Images

* Adding a new section: Removing Unwanted FITS Files

* Adding a new section: Filtering FITS Files by Exposure Time

* Reviewing code, markdown text, and comments

* Reviewing code, markdown, and doc-strings.

*
[…]
