+++ requirements/docs.txt
@@ -27,3 +27,5 @@ sphinx_rtd_theme >= 1.2
 sphinxcontrib-bibtex >= 2.5
 sunpy[all]
 towncrier >= 22.12
+imageio
+ipywidgets
