+++ setup.cfg
@@ -78,6 +78,8 @@ docs =
     sphinxcontrib-bibtex >= 2.5
     sunpy[net] >= 3.0.0
     towncrier >= 22.12
+    imageio
+    ipywidgets
 developer =
     # install everything for developers
     # ought to functionally mirror requirements.txt
