### 2024-06-05 2369b55
Create documentation stub files (#254)
* Update api_static pages

* Add imports

* Restructuring to get docs to build

* Use SSW_MIRROS from xrtpy.util

* More consistent naming

* Fix namespace

* Update __all__ defs

* pre-commit run --all-files
