+++ docs/api_static/xrtpy.image_correction.deconvolve.rst
@@ -1,0 +1,8 @@
+:orphan:
+
+`xrtpy.image_correction.deconvolve`
+===================================
+
+.. currentmodule:: xrtpy.image_correction.deconvolve
+
+.. automodapi:: xrtpy.image_correction.deconvolve
