+++ docs/api_static/xrtpy.util.time.rst
@@ -1,0 +1,8 @@
+:orphan:
+
+`xrtpy.util.time`
+=================
+
+.. currentmodule:: xrtpy.util.time
+
+.. automodapi:: xrtpy.util.time
