### 2024-06-05 008e708
Update xrtpy/response/effective_area.py
Co-authored-by: Nick Murphy <namurphy@cfa.harvard.edu>
