### 2024-06-05 77f2df1
Update xrtpy/response/effective_area.py
Co-authored-by: Nick Murphy <namurphy@cfa.harvard.edu>
