### 2024-06-05 dbbe9d6
Update xrtpy/response/effective_area.py
Co-authored-by: Nick Murphy <namurphy@cfa.harvard.edu>
