### 2024-06-05 a2f61db
Functionality documentation Update - Effective Area  (#250)
* Updated two functions

* Leaving contamination_on_CCD since intended for external use

* Adding docstring to effective_area function

* updated some doc strings and covert function to be used internally

* Fixing Linters

* Coverting function to be used internally

* Coverting function to be used internally

* Coverting function to be used internally-_filterwheel_angular_wavenumber

* Coverting function to be used internally- _angular_wavenumber_CCD

* Coverting function to be used internally- _CCD_contamination_transmission

* Cove
[…]
