+++ xrtpy/response/effective_area.py
@@ -3,7 +3,6 @@ __all__ = [
     "effective_area",
 ]
 
-
 import datetime
 import math
 import os
@@ -66,16 +65,19 @@ _filter_contamination = _filter_contam_file["p2"]
 
 class EffectiveAreaFundamental:
     """
-    Class for calculating the effective area.
+    Class for calculating the effective area for an XRT filter at a specific observation date.
+
+    This class handles the calculations required to determine the effective area of a filter
+    used in the X-Ray Telescope (XRT) on the Hinode satellite. It considers various factors such as
+    contamination on the CCD and filters, as well as the geometry and transmission of the XRT channel.
 
     Parameters
-    -----------
+    ----------
     filter_name : str
         The name of the filter.
-
-    observation_date: str
-        The date of the observation.  For valid date formats, look at the documentation for
-        `sunpy.time.parse_time`.
+    observation_date : str or datetime.datetime
+        The date of the observation. Acceptable formats include any string or datetime object
+        that can be parsed by `sunpy.time.parse_time`.
     """
 
     def __init__(self, filter_name, observation_date):
@@ -85,12 +87,20 @@ class EffectiveAreaFundamental:
 
     @property
     def name(self) -> str:
-        """Name of XRT X-Ray channel filter."""
+        """
+        Name of XRT X-Ray channel filter.
+
+        :noindex:
+        """
         return self._name
 
     @property
     def observation_date(self) -> str:
-        """Date of observation."""
+        """
+        Date of observation.
+
+        :noindex:
+        """
         return self._observation_date
 
     @observation_date.setter
@@ -129,14 +139,36 @@ class EffectiveAreaFundamental:
 
     @property
     def contamination_on_CCD(self):
-        """Calculation of contamination layer on the CCD, thickness given in Angstrom (Å)."""
+        """
+        Calculate the thickness of the contamination layer on the CCD.
+
+        This property interpolates the contamination data over time to determine the thickness
+        of the contamination layer on the CCD at the observation date. The contamination layer
+        is measured in Angstroms (Å).
+
+        Returns
+        -------
+        astropy.units.Quantity
+            The thickness of the contamination layer on the CCD, in Angstroms.
+
+        Notes
+        -----
+        The interpolation is performed using a linear interpolation method over the
+        available contamination data points. The `observation_date` attribute is used to
+        provide the point at which to evaluate the interpolation.
+
+        Raises
+        ------
+        ValueError
+            If the observation date is outside the range of the available contamination data.
+        """
         interpolater = scipy.interpolate.interp1d(
             _ccd_contamination_file_time.utime, _ccd_contamination, kind="linear"
         )
         return interpolater(self.observation_date.utime)
 
     @property
-    def filter_index_mapping_to_name(self):
+    def _filter_index_mapping_to_name(self):
         """Returns filter's corresponding number value."""
         if self.name in index_mapping_to_fw1_name:
             return index_mapping_to_fw1_name.get(self.name)
@@ -144,33 +176,33 @@ class EffectiveAreaFundamental:
             return index_mapping_to_fw2_name.get(self.name)
 
     @property
-    def filter_wheel_number(self):
+    def _filter_wheel_number(self):
         """Defining chosen filter to its corresponding filter wheel."""
         return 0 if self.name in index_mapping_to_fw1_name else 1
 
     @property
-    def combo_filter_name_split(self):
+    def _combo_filter_name_split(self):
         """Defining chosen filters to its corresponding filter wheel."""
         name = (self.name).split("/")
         filter1, filter2 = name[0], name[1]
         return filter1, filter2
 
     @property
[…]
