### 2024-06-05 #6396 @joyvelasquez
Hi @dfm, the XRTpy team has made all the edits suggested by the reviewers, and all changes have been merged into
XRTpy. We are planning to release a update version of the package with all the JOSS-related changes and additional new features by the end of the month. I believe the XRTpy JOSS paper is ready for another review. How should we proceed from here?
