+++ docs/changelog/0.5.0.rst
@@ -3,6 +3,7 @@ XRTpy v0.5.0 (2024-06-21)
 
 New Features
 ------------
+- New Example Notebook: Introduced a new example notebook titled "Hinode-XRT: A Practical Guide to Data Extraction and Visualization." This notebook provides a step-by-step guide on how to download, process, and visualize solar observation data from the Hinode X-Ray Telescope, making it accessible for research and analysis.(:pr:`210`)
 
 Deprecations and Removals
 -------------------------
