### 2024-06-21 a71f070
Adding updated IDL EA test file
