+++ docs/changelog/0.5.0.rst
@@ -14,3 +14,4 @@ Documentation
 Internal modifications and improvements
 ---------------------------------------
 - Updated the `xrt_contam_on_ccd.geny` file with new CCD contamination values. (:pr:`245`)
+- Updated IDL result - Effective Area testing files for the following years: 2020,2021,2022,2023,2024. (:pr:`218`)
