+++ docs/changelog/0.5.0.rst
@@ -3,7 +3,7 @@ XRTpy v0.5.0 (2024-06-21)
 
 New Features
 ------------
-- New Example Notebook: Introduced a new example notebook titled "Hinode-XRT: A Practical Guide to Data Extraction and Visualization." This notebook provides a step-by-step guide on how to download, process, and visualize solar observation data from the Hinode X-Ray Telescope, making it accessible for research and analysis.(:pr:`210`)
+- New Example Notebook: Introduced a new example notebook titled "Hinode-XRT: A Practical Guide to Data Extraction and Visualization." This notebook provides a step-by-step guide on how to download, process, and visualize solar observation data from the Hinode X-Ray Telescope, making it accessible for research and analysis. (:pr:`210`)
 
 Documentation Updates
 ---------------------
@@ -11,7 +11,7 @@ Documentation Updates
 
 Internal modifications and improvements
 ---------------------------------------
-- CCD Contamination Update: Updated the xrt_contam_on_ccd.geny file with new CCD contamination values. (:pr:`245`)
+- CCD Contamination Update: Updated the `xrt_contam_on_ccd.geny` file with new CCD contamination values. (:pr:`245`)
 - Effective Area Testing Files Update: Updated IDL result - Effective Area testing files for the following years: 2020, 2021, 2022, 2023, 2024. (:pr:`218`)
 - Temperature Response Testing Files Update: Updated IDL result - Temperature Response testing files for the following years: 2020, 2021, 2022, 2023, 2024. (:pr:`220`)
 - SunPy Maps Import Method: Modified import methods for SunPy maps in the XRTpy notebook to align with the way it's imported in other notebooks for consistency. (:pr:`237`)
