### 2024-06-21 d6900b4
Update docs/changelog/0.5.0.rst
Co-authored-by: Nick Murphy <namurphy@cfa.harvard.edu>
