+++ docs/changelog/0.5.0.rst
@@ -7,7 +7,7 @@ New Features
 
 Documentation Updates
 ---------------------
-- Contributing Information: Updated informational methods for contributing to xrtpy through different avenues such as reporting issues, submitting pull requests, or improving documentation. (:pr:`248`)
+- Contributing Information: Updated information for contributing to XRTpy through different avenues such as reporting issues, submitting pull requests, or improving documentation. (:pr:`248`)
 
 Internal modifications and improvements
 ---------------------------------------
