### 2024-06-21 31a4612
Update docs/changelog/0.5.0.rst
Co-authored-by: Nick Murphy <namurphy@cfa.harvard.edu>
