+++ docs/changelog/0.5.0.rst
@@ -14,7 +14,7 @@ Internal modifications and improvements
 - CCD Contamination Update: Updated the :file:`xrt_contam_on_ccd.geny` file with new CCD contamination values. (:pr:`245`)
 - Effective Area Testing Files Update: Updated IDL result - Effective Area testing files for the following years: 2020, 2021, 2022, 2023, 2024. (:pr:`218`)
 - Temperature Response Testing Files Update: Updated IDL result - Temperature Response testing files for the following years: 2020, 2021, 2022, 2023, 2024. (:pr:`220`)
-- SunPy Maps Import Method: Modified import methods for SunPy maps in the XRTpy notebook to align with the way it's imported in other notebooks for consistency. (:pr:`237`)
+- SunPy Maps Import Method: Modified import methods for SunPy `~sunpy.map.Map` objects in the XRTpy notebook for consistency with the way it is imported in other notebooks. (:pr:`237`)
 - Complete Review of Function Documentation: Conducted a complete review of documentation for every function in the following XRTpy objects: Channel, Effective Area, and Temperature Response. (:pr:`253,250,249`)
 
 Miscellaneous
