### 2024-06-21 c8bff5b
Update docs/changelog/0.5.0.rst
Co-authored-by: Nick Murphy <namurphy@cfa.harvard.edu>
