### 2024-06-21 b4e9d35
Updating Effective Area Script - _interpolated_CCD_contamination_transmission
