### 2024-06-21 1bb67ef
Updating effective area function in the temperature response script
