### 2024-06-21 70473a5
Updating Effective Area Script - _interpolated_CCD_contamination_transmission
