### 2024-06-21 2d2344f
Updating effective area function in the temperature response script
