### 2024-06-21 c673f11
Updating to OG function
