### 2024-06-21 9c19fa5
Updating to OG function
