### 2024-06-24 #6396 @Cadair
Hi all,

I think it looks good to me. My remaining comment about the docs would be that I would like to see some kind of index of all the API to make it easier to find, but it is there if you click through so I am happy to tick all the boxes for the review.
