### 2024-06-24 83cdc8e
Convert setup.cfg to pyproject.toml
