### 2024-06-24 7ff6ccf
Make filename of version.py consistent with past usage
