### 2024-06-24 39ccd74
Expand & update requirements
