### 2024-06-24 cfcc6b9
Remove packaging as a dependency
