+++ pyproject.toml
@@ -43,7 +43,6 @@ dependencies = [
   "cached-property >= 1.5.2",
   "matplotlib >= 3.5.0",
   "numpy >= 1.24.0",
-  "packaging",
   "scipy >= 1.8.0",
   "sunpy[map] >= 4.0.0",
   "scikit-image >= 0.19.0",
