### 2024-06-24 d437a1f
Remove developer set of dependencies
