+++ pyproject.toml
@@ -95,40 +95,6 @@ docs = [
   "imageio",
   "ipywidgets",
 ]
-developer = [
-  # install everything for developers
-  # ought to functionally mirror requirements.txt
-  "codespell",
-  "pre-commit",
-  "nox",
-  "docutils >= 0.18",
-  "ipykernel",
-  "ipython",
-  "jinja2 != 3.1",
-  "nbconvert < 7.14",
-  "nbsphinx >= 0.9",
-  "numpydoc >= 1.2",
-  "pillow",
-  "pygments >= 2.11",
-  "sphinx >= 6.0",
-  "sphinx-changelog >= 1.3",
-  "sphinx-codeautolink >= 0.15.0",
-  "sphinx-copybutton >= 0.5",
-  "sphinx-gallery >= 0.12",
-  "sphinx-hoverxref >= 1.1",
-  "sphinx-issues >= 3.0",
-  "sphinx_automodapi",
-  "sphinx_rtd_theme >= 1.2",
-  "sphinxcontrib-bibtex >= 2.5",
-  "sunpy[net] >= 3.0.0",
-  "towncrier >= 22.12",
-  "imageio",
-  "ipywidgets",
-  "pydocstyle",
-  "pytest >= 5.4.0",
-  "pytest-allclose",
-  "pytest-xdist",
-]
 
 [project.urls]
 Documentation = "https://xrtpy.readthedocs.io"
