+++ pyproject.toml
@@ -4,7 +4,7 @@ requires = [
   "setuptools>=50",
   "setuptools_scm>=6",
   "wheel>=0.34",
-]  # ought to mirror 'requirements/build.txt'
+]
 
 [project]
 name = "xrtpy"
@@ -43,42 +43,33 @@ dependencies = [
   "cached-property >= 1.5.2",
   "matplotlib >= 3.5.0",
   "numpy >= 1.24.0",
+  "requests >= 2.28.0",
+  "scikit-image >= 0.19.0",
   "scipy >= 1.8.0",
   "sunpy[map] >= 4.0.0",
-  "scikit-image >= 0.19.0",
-  "requests >= 2.28.0",
 ]
 
 [project.optional-dependencies]
-extras = [
-  # ought to mirror requirements/extras.txt
-  "codespell",
-  "pre-commit",
-  "nox",
-]
 tests = [
-  # ought to mirror requirements/tests.txt
-  "codespell",
-  "pre-commit",
   "nox",
-  "pydocstyle",
+  "pre-commit",
   "pytest >= 5.4.0",
   "pytest-allclose",
   "pytest-xdist",
 ]
 docs = [
-  # ought to mirror requirements/docs.txt
-  "codespell",
-  "pre-commit",
-  "nox",
   "docutils >= 0.18",
+  "imageio",
   "ipykernel",
   "ipython",
+  "ipywidgets",
   "jinja2 != 3.1",
   "nbconvert < 7.14",
   "nbsphinx >= 0.9",
+  "nox",
   "numpydoc >= 1.2",
   "pillow",
+  "pre-commit",
   "pygments >= 2.11",
   "sphinx >= 6.0",
   "sphinx-changelog >= 1.3",
@@ -90,10 +81,8 @@ docs = [
   "sphinx_automodapi",
   "sphinx_rtd_theme >= 1.2",
   "sphinxcontrib-bibtex >= 2.5",
-  "sunpy[net] >= 3.0.0",
+  "sunpy[net] >= 4.0.0",
   "towncrier >= 22.12",
-  "imageio",
-  "ipywidgets",
 ]
 
 [project.urls]
