### 2024-06-24 c8d39f4
Update requirements & usage for tests/docs/dev
