+++ noxfile.py
@@ -21,7 +21,7 @@ pytest_options = [
 
 @nox.session(python=python_versions)
 def tests(session):
-    session.install(".[tests]")
+    session.install(".[dev,tests]")
     session.run("pytest", *pytest_options)
 
 
@@ -39,7 +39,7 @@ def import_package(session):
 
 @nox.session
 def build_docs(session):
-    session.install(".[docs]")
+    session.install(".[dev,docs]")
     session.run(
         "sphinx-build",
         *sphinx_opts,
@@ -49,8 +49,7 @@ def build_docs(session):
 
 @nox.session
 def build_docs_nitpicky(session):
-    session.install("-r", "requirements/docs.txt")
-    session.install(".")
+    session.install(".[dev,docs]")
     session.run(
         "sphinx-build",
         *sphinx_opts,
@@ -61,8 +60,7 @@ def build_docs_nitpicky(session):
 
 @nox.session
 def build_docs_no_examples(session):
-    session.install("-r", "requirements/docs.txt")
-    session.install(".")
+    session.install(".[dev,docs]")
     session.run(
         "sphinx-build",
         *sphinx_opts,
