+++ pyproject.toml
@@ -50,39 +50,39 @@ dependencies = [
 ]
 
 [project.optional-dependencies]
+dev = [
+  "nox >= 2022.8.7",
+  "pre-commit >= 3.6.0",
+]
 tests = [
-  "nox",
-  "pre-commit",
-  "pytest >= 5.4.0",
-  "pytest-allclose",
-  "pytest-xdist",
+  "pytest >= 8.0.0",
+  "pytest-allclose >= 1.0.0",
+  "pytest-xdist >= 3.6.1",
 ]
 docs = [
-  "docutils >= 0.18",
-  "imageio",
-  "ipykernel",
-  "ipython",
-  "ipywidgets",
-  "jinja2 != 3.1",
-  "nbconvert < 7.14",
+  "docutils >= 0.21.2",
+  "imageio >= 2.20.0",
+  "ipykernel >= 6.20.0",
+  "ipython >= 8.4.0",
+  "ipywidgets >= 8.8.0",
+  "jinja2 != 3.1, >= 3.0.0",
+  "nbconvert >= 7.7.0, < 7.14",
   "nbsphinx >= 0.9",
-  "nox",
-  "numpydoc >= 1.2",
-  "pillow",
-  "pre-commit",
-  "pygments >= 2.11",
-  "sphinx >= 6.0",
-  "sphinx-changelog >= 1.3",
-  "sphinx-codeautolink >= 0.15.0",
-  "sphinx-copybutton >= 0.5",
-  "sphinx-gallery >= 0.12",
-  "sphinx-hoverxref >= 1.1",
-  "sphinx-issues >= 3.0",
-  "sphinx_automodapi",
-  "sphinx_rtd_theme >= 1.2",
-  "sphinxcontrib-bibtex >= 2.5",
+  "numpydoc >= 1.5.0",
+  "pillow >= 9.1.0",
+  "pygments >= 2.12.0",
+  "sphinx >= 7.3.0",
+  "sphinx-changelog >= 1.5.0",
+  "sphinx-codeautolink >= 0.15.2",
+  "sphinx-copybutton >= 0.5.2",
+  "sphinx-gallery >= 0.16.0",
+  "sphinx-hoverxref >= 1.4.0",
+  "sphinx-issues >= 4.1.0",
+  "sphinx_automodapi >= 0.17.0",
+  "sphinx_rtd_theme >= 2.0.0",
+  "sphinxcontrib-bibtex >= 2.6.2",
   "sunpy[net] >= 4.0.0",
-  "towncrier >= 22.12",
+  "towncrier >= 23.11.0",
 ]
 
 [project.urls]
