### 2024-06-25 5340537
Fix ipywidgets dependency
