+++ pyproject.toml
@@ -64,7 +64,7 @@ docs = [
   "imageio >= 2.20.0",
   "ipykernel >= 6.20.0",
   "ipython >= 8.4.0",
-  "ipywidgets >= 8.8.0",
+  "ipywidgets >= 8.1.0",
   "jinja2 != 3.1, >= 3.0.0",
   "nbconvert >= 7.7.0, < 7.14",
   "nbsphinx >= 0.9",
