### 2024-06-25 b6838f0
Install setuptools for doc builds to get pkg_resources
