+++ pyproject.toml
@@ -71,6 +71,8 @@ docs = [
   "numpydoc >= 1.5.0",
   "pillow >= 9.1.0",
   "pygments >= 2.12.0",
+  # install setuptools to get pkg_resources for doc build
+  "setuptools; python_version >= '3.12'",
   "sphinx >= 7.3.0",
   "sphinx-changelog >= 1.5.0",
   "sphinx-codeautolink >= 0.15.2",
