### 2024-06-25 ab0cd76
Stop separting initial-tests from comprehensive tests
