### 2024-06-27 65ea019
Require setuptools for Python ≥ 3.12 to get pkg_resources
