+++ pyproject.toml
@@ -46,6 +46,8 @@ dependencies = [
   "requests >= 2.28.0",
   "scikit-image >= 0.19.0",
   "scipy >= 1.8.0",
+  # install setuptools to get pkg_resources
+  "setuptools; python_version >= '3.12'",
   "sunpy[map] >= 4.0.0",
 ]
 
@@ -71,8 +73,6 @@ docs = [
   "numpydoc >= 1.5.0",
   "pillow >= 9.1.0",
   "pygments >= 2.12.0",
-  # install setuptools to get pkg_resources for doc build
-  "setuptools; python_version >= '3.12'",
   "sphinx >= 7.3.0",
   "sphinx-changelog >= 1.5.0",
   "sphinx-codeautolink >= 0.15.2",
