### 2024-06-28 406107c
Update nox session for running tests
