+++ noxfile.py
@@ -1,6 +1,6 @@
 import nox
 
-nox.options.sessions = ["tests", "linters"]
+nox.options.sessions = ["tests"]
 
 python_versions = ("3.9", "3.10", "3.11", "3.12")
 
@@ -19,11 +19,11 @@ pytest_options = [
 ]
 
 
-@nox.session(python=python_versions)
+@nox.session
 def tests(session):
     session.install("-r", "requirements/tests.txt")
     session.install(".")
-    session.run("pytest", *pytest_options)
+    session.run("pytest", *pytest_options, *session.posargs)
 
 
 @nox.session
