### 2024-06-28 9c0cbdc
Update nox session for running tests
