### 2024-06-28 d7d18c3
Convert `setup.cfg` to `pyproject.toml` (#260)
* Convert setup.cfg to pyproject.toml

* Remove setup.py

* Make filename of version.py consistent with past usage

* Remove extra blank line

* Expand & update requirements

* Remove packaging as a dependency

* Remove developer set of dependencies

* Alphabetize requirements

* Stop using requirements files for Nox sessions

* Remove requirements files as no longer needed

* Update requirements & usage for tests/docs/dev

* Fix ipywidgets dependency

* Install setuptools for doc builds to get pkg_resources

* Relax requirement on docutils

* Req
[…]
