+++ pyproject.toml
@@ -4,7 +4,94 @@ requires = [
   "setuptools>=50",
   "setuptools_scm>=6",
   "wheel>=0.34",
-]  # ought to mirror 'requirements/build.txt'
+]
+
+[project]
+name = "xrtpy"
+readme = "README.md"
+keywords = ["solar physics"]
+description = "For analyzing data from the X-Ray Telescope (XRT) on the Hinode spacecraft."
+license = {file = "LICENSE"}
+classifiers = [
+  "Development Status :: 4 - Beta",
+  "Intended Audience :: Science/Research",
+  "License :: OSI Approved :: BSD License",
+  "Operating System :: OS Independent",
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3.9",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: Implementation :: CPython",
+  "Topic :: Scientific/Engineering :: Astronomy",
+  "Topic :: Scientific/Engineering :: Physics",
+]
+# SPEC 0 recommends that packages in the scientific pythoniverse support
+# versions of Python that have been released in the last ≤36 months
+requires-python = ">=3.9"
+dynamic = ["version"]
+authors = [
+  {name = "Joy Velasquez", email = "joy.velasquez@cfa.harvard.edu"},
+  {name = "Jonathan Slavin", email = "jslavin@cfa.harvard.edu"},
+  {name = "Nick Murpy", email="namurphy@cfa.harvard.edu"},
+  {name = "Will Barnes"},
+  {name = "Stuart Mumford"},
+]
+
+dependencies = [
+  "astropy >= 5.1",
+  "cached-property >= 1.5.2",
+  "matplotlib >= 3.5.0",
+  "numpy >= 1.24.0",
+  "requests >= 2.28.0",
+  "scikit-image >= 0.19.0",
+  "scipy >= 1.8.0",
+  # install setuptools to get pkg_resources
+  "setuptools; python_version >= '3.12'",
+  "sunpy[map] >= 4.0.0",
+]
+
+[project.optional-dependencies]
+dev = [
+  "nox >= 2022.8.7",
+  "pre-commit >= 3.6.0",
+]
+tests = [
+  "pytest >= 8.0.0",
+  "pytest-allclose >= 1.0.0",
+  "pytest-xdist >= 3.6.1",
+]
+docs = [
+  "docutils >= 0.19",
+  "imageio >= 2.20.0",
+  "ipykernel >= 6.20.0",
+  "ipython >= 8.4.0",
+  "ipywidgets >= 8.1.0",
+  "jinja2 != 3.1, >= 3.0.0",
+  "nbconvert >= 7.7.0, < 7.14",
+  "nbsphinx >= 0.9",
+  "numpydoc >= 1.5.0",
+  "pillow >= 9.1.0",
+  "pygments >= 2.12.0",
+  "sphinx >= 7.3.0",
+  "sphinx-changelog >= 1.5.0",
+  "sphinx-codeautolink >= 0.15.2",
+  "sphinx-copybutton >= 0.5.2",
+  "sphinx-gallery >= 0.16.0",
+  "sphinx-hoverxref >= 1.4.0",
+  "sphinx-issues >= 4.1.0",
+  "sphinx_automodapi >= 0.17.0",
+  "sphinx_rtd_theme >= 2.0.0",
+  "sphinxcontrib-bibtex >= 2.6.2",
+  "sunpy[net] >= 4.0.0",
+  "towncrier >= 23.11.0",
+]
+
+[project.urls]
[…]
