### 2024-06-28 a2dbe58
Remove extraneous Nox sessions for docs
