### 2024-06-28 0eb93a3
Remove extraneous Nox sessions for docs
