### 2024-06-28 ae4a5a6
Finish renaming nox sessions
