### 2024-06-28 1b14275
Finish renaming nox sessions
