### 2024-06-28 a3607f9
Finish consolidating Nox sessions
