+++ noxfile.py
@@ -1,44 +1,50 @@
 import nox
 
 nox.options.sessions = ["tests", "linters"]
+nox.options.default_venv_backend = "uv|virtualenv"
 
 python_versions = ("3.9", "3.10", "3.11", "3.12")
-
-sphinx_paths = ["docs", "docs/_build/html"]
-sphinx_fail_on_warnings = ["-W", "--keep-going"]
-sphinx_builder = ["-b", "html"]
-sphinx_opts = sphinx_paths + sphinx_fail_on_warnings + sphinx_builder
-sphinx_no_notebooks = ["-D", "nbsphinx_execute=never"]
-sphinx_nitpicky = ["-n"]
-
-pytest_options = [
-    "--ignore",
-    "xrtpy/response/effective_area.py",
-    "--ignore",
-    "xrtpy/response/temperature_response.py",
-]
 
 
 @nox.session(python=python_versions)
 def tests(session):
+    """Run tests with pytest."""
+
+    pytest_options = [
+        "--ignore",
+        "xrtpy/response/effective_area.py",
+        "--ignore",
+        "xrtpy/response/temperature_response.py",
+    ]
+
     session.install(".[dev,tests]")
     session.run("pytest", *pytest_options, *session.posargs)
 
 
 @nox.session
 def linters(session):
+    """Run all pre-commit hooks on all files."""
     session.install("pre-commit")
     session.run("pre-commit", "run", "--all-files", *session.posargs)
 
 
 @nox.session
 def import_package(session):
+    """Import xrtpy."""
     session.install(".")
     session.run("python", "-c", 'import xrtpy')  # fmt: skip
 
 
 @nox.session
 def docs(session):
+    """Build documentation with Sphinx."""
+
+    sphinx_paths = ["docs", "docs/_build/html"]
+    sphinx_fail_on_warnings = ["-W", "--keep-going"]
+    sphinx_builder = ["-b", "html"]
+    sphinx_nitpicky = ["-n"]
+    sphinx_opts = sphinx_paths + sphinx_fail_on_warnings + sphinx_builder + sphinx_nitpicky
+
     session.install(".[dev,docs]")
     session.run(
         "sphinx-build",
