### 2024-07-20 06d2917
allow for extrapolation in CCD contaminination
