### 2024-07-28 #6396 @dfm
@joyvelasquez — Thanks for checking in! I've gone back to @dpshelio's [comment above](https://github.com/openjournals/joss-reviews/issues/6396#issuecomment-2026099021) and it looks like it should be fine to continue without a full checklist. I'm sorry for missing that!

I have a few last processing steps for you to do now. Can you first update the two references to `\ssr` in the biblography file:

```diff
-   journal = {\ssr},
+   journal = {Space Science Reviews},
```

Once you've done that:

1. Comment `@editorialbot generate pdf` on this thread and take one last read through the manuscript to make sure that you're happy with it (it's harder to make changes later!), especially the author names and affiliations. I've taken a pass and it looks good to me!
