### 2024-07-29 0e598f9
Remove .py from class name reference in joss paper.
