### 2024-07-29 dc4bc7c
Updated temperature response plot reference comment
