### 2024-07-29 5d8194a
Updated effective area plot reference comment
