### 2024-07-29 749244d
Final Updates for XRTpy JOSS Submission (#267)
* Update the two references -ssr

* Remove .py from class name reference in joss paper.

* Updated effective area plot-fine detail

* Updated temperature response-fine detail

* Updated temperature response plot reference comment

* Updated effective area plot reference comment

* Updated effective area units

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
