### 2024-07-30 f947177
Updating text info about XRT and adding image
