+++ docs/about_xrt.rst
@@ -25,8 +25,18 @@ September 23, 2006 (Japan Standard Time) and placed into a polar,  sun-synchrono
 
 The X-Ray Telescope
 ====================
-The X-Ray Telescope (XRT) is an instrument onboard the Hinode spacecraft. With its remarkable capabilities, the XRT provides high-resolution
-images of the solar corona's hottest material, ranging from 1,000,000 to 10,000,000 Kelvin. Visit the official `xrt-cfa-harvard`_ website for mission overview information about the XRT.
+The X-Ray Telescope (XRT), depicted as a long linear black tube on the Hinode spacecraft (illustration at the top of the page ),
+ is a crucial instrument for observing the solar corona's most intense regions, with temperatures ranging from 1,000,000 to 10,000,000
+ Kelvin. The image on the bellow is a synoptic composite from February 14, 2015, created using the Al-Mesh/Be-Thin/Al-Med filters.
+ This method combines images taken with different filters, each set to a distinct color in the RGB color model to highlight various
+ thermal conditions within the corona. For a comprehensive overview of XRT's mission and capabilities, please visit the
+  official `xrt-cfa-harvard`_ website for mission overview information about the XRT.
+
+.. image:: images/XRT_composite_image_full_disk_14February2015.png
+   :alt: XRT Composite Image
+   :align: center
+   :scale: 64%
+
 
 .. tip::
    Visit the `XRT Picture of the Week`_ and the `Hinode-XRT YouTube`_ page for captivating visual content showcasing the XRT's solar observations.
