### 2024-07-30 73c0022
Updating text info about XRT
