### 2024-07-31 2cc6ed8
Updating text info about XRT-part 2
