### 2024-07-31 f329912
Updated summary info about XRT Obs.
