### 2024-07-31 ff713a3
Adding Dere:1997 reference - Chianti
