+++ joss/paper.bib
@@ -10,6 +10,19 @@
   note = {Volume publication date: 22 September 2009}
 }
 
+@ARTICLE{Dere:1997,
+  author = {K. P. Dere and E. Landi and H.E. Mason and B. C. Monsignori Fossi and P. R. Young},
+  title = "{CHIANTI - An Atomic Database For Emission Lines - Paper I: Wavelengths greater than 50 Å}",
+  journal = {Astronomy and Astrophysics Suppl. Ser.},
+  keywords = {},
+  year = {1997},
+  month = {},
+  volume = {125},
+  pages = {149-173},
+  adsurl = {https://ui.adsabs.harvard.edu/abs/1997A&AS..125..149D/abstract},
+  adsnote = {Available from ADS}
+}
+
 @ARTICLE{Fludra-and-Schmelz:1999,
   author = {{Fludra}, A. and {Schmelz}, J.~T.},
   title = "{The absolute coronal abundances of sulfur, calcium, and iron from
