### 2024-07-31 75d2ea2
Adding DelZanna:2020 reference - Chianti V10
