+++ joss/paper.bib
@@ -23,6 +23,19 @@
   adsnote = {Available from ADS}
 }
 
+@ARTICLE{DelZanna:2020,
+  author = {G. Del Zanna and K.P. Dere and P.R. Young and E. Landi},
+  title = "{CHIANTI - An Atomic Database for Emission Lines. Paper 16, Version 10.0, Further Extensions}",
+  journal = {The Astrophysical Journal},
+  keywords = {},
+  year = {2021},
+  month = {},
+  volume = {909},
+  pages = {38},
+  adsurl = {https://ui.adsabs.harvard.edu/abs/2021ApJ...909...38D/abstract},
+  adsnote = {Available from ADS}
+}
+
 @ARTICLE{Fludra-and-Schmelz:1999,
   author = {{Fludra}, A. and {Schmelz}, J.~T.},
   title = "{The absolute coronal abundances of sulfur, calcium, and iron from
