### 2024-07-31 a43ccd8
Adding Chianti reference
