+++ joss/paper.md
@@ -75,7 +75,7 @@ The package is structured into distinct modules, each serving a specific purpose
 
  - `xrtpy.response.effective_area`: XRTpy's capability to calculate effective areas for various XRT filter channels, combined with CCD contamination layer thickness information, is crucial for understanding instrumental spectral responses, as depicted in \autoref{fig:Figure 1}.
 
- - `xrtpy.response.temperature_response`: XRTpy provides the capability of computing the temperature response of all the XRT filter channels. It does this by relying on a spectral emission model, drawing from [@Narukage:2011] and [@Narukage:2014]. Users can choose from a range of CHIANTI abundance sets, including the default model with coronal abundances, hybrid, and photospheric options. \autoref{fig:Figure 2} shows XRTpy's temperature response calculations for all XRT filters across the different CHIANTI  abundance sets.
+ - `xrtpy.response.temperature_response`: XRTpy provides the capability of computing the temperature response of all the XRT filter channels. It does this by relying on a spectral emission model, drawing from [@Narukage:2011] and [@Narukage:2014]. Users can choose from a range of CHIANTI abundance sets, including the default model with coronal abundances, hybrid, and photospheric options. \autoref{fig:Figure 2} shows XRTpy's temperature response calculations for all XRT filters across the different CHIANTI abundance sets. The CHIANTI database is described in [@Dere:1997] and the version used in XRTpy is CHIANTI version 10.0 [@Del_Zanna:2020].
 
  - `xrtpy.response.temperature_from_filter_ratio`: This module contains the `temperature_from_filter_ratio` function, which derives temperature and emission measure maps for a pair of images using the filter ratio method. \autoref{fig:Figure 3} illustrates an example usage of this function.
 
