### 2024-07-31 55bc650
Moved setence
