### 2024-07-31 40f206d
Adding log text to a plot label
