### 2024-07-31 2edea53
Removing Figure 3-repeated
