### 2024-07-31 46a087b
Adding Weberg EISpack paper
