+++ joss/paper.bib
@@ -205,3 +205,16 @@ Data }",
   publisher = {},
   url = {https://ui.adsabs.harvard.edu/abs/2020ApJ...890...68S/abstract},
 }
+
+@article{Weberg2023,
+  doi = {10.21105/joss.04914},
+  url = {https://doi.org/10.21105/joss.04914},
+  year = {2023},
+  publisher = {The Open Journal},
+  volume = {8},
+  number = {85},
+  pages = {4914},
+  author = {Micah J. Weberg and Harry P. Warren and Nicholas Crump and Will Barnes},
+  title = {EISPAC - The EIS Python Analysis Code},
+  journal = {Journal of Open Source Software}
+}
