### 2024-07-31 8d342f0
Adding Weberg EISpack paper-forgot to add :
