### 2024-07-31 3b86e83
Adding Weberg EISpack paper-forgot to add :
