+++ joss/paper.bib
@@ -206,7 +206,7 @@ Data }",
   url = {https://ui.adsabs.harvard.edu/abs/2020ApJ...890...68S/abstract},
 }
 
-@article{Weberg2023,
+@article{Weberg:2023,
   doi = {10.21105/joss.04914},
   url = {https://doi.org/10.21105/joss.04914},
   year = {2023},
