+++ joss/paper.bib
@@ -10,6 +10,19 @@
   note = {Volume publication date: 22 September 2009}
 }
 
+@article{Barnes:2020,
+  doi = {10.21105/joss.02801},
+  url = {https://doi.org/10.21105/joss.02801},
+  year = {2020},
+  publisher = {The Open Journal},
+  volume = {5},
+  number = {55},
+  pages = {2801},
+  author = {Will T. Barnes and Mark C. m. Cheung and Monica G. Bobra and Paul F. Boerner and Georgios Chintzoglou and Drew Leonard and Stuart J. Mumford and Nicholas Padmanabhan and Albert Y. Shih and Nina Shirman and David Stansby and Paul J. Wright},
+  title = {aiapy: A Python Package for Analyzing Solar EUV Image Data from AIA},
+  journal = {Journal of Open Source Software}
+}
+
 @ARTICLE{Dere:1997,
   author = {K. P. Dere and E. Landi and H.E. Mason and B. C. Monsignori Fossi and P. R. Young},
   title = "{CHIANTI - An Atomic Database For Emission Lines - Paper I: Wavelengths greater than 50 Å}",
