+++ joss/paper.bib
@@ -207,18 +207,15 @@ Data }",
 
 @article{SunPy:2020,
   title = {The SunPy Project: Open Source Development and Status of the Version 1.0 Core Package},
-  author = {SunPy Community; Barnes, Will T.; Bobra, Monica G.; Christe, Steven D.; Freij, Nabil; Hayes, Laura A.; Ireland, Jack; Mumford, Stuart; Perez-Suarez, David; Ryan, Daniel F.; Shih, Albert Y.; and others},
+  author = {SunPy Community and Will T. Barnes and Monica G. Bobra and Steven D. Christe and Nabil Freij and Laura A. Hayes and Jack Ireland and Stuart Mumford and David Perez-Suarez and Daniel F. Ryan and Albert Y. Shih and (Primary Paper Contributors) and Prateek Chanda and Kolja Glogowski and Russell Hewett and V. Keith Hughitt and Andrew Hill and Kaustubh Hiware and Andrew Inglis and Michael S. F. Kirk and Sudarshan Konge and James Paul Mason and Shane Anthony Maloney and Sophie A. Murray and Asish Panda and Jongyeob Park and Tiago M. D. Pereira and Kevin Reardon and Sabrina Savage and Brigitta M. Sipőcz and David Stansby and Yash Jain and Garrison Taylor and Tannmay Yadav and Rajul and Trung Kien Dang and (Sunpy Contributors)},
   journal = {The Astrophysical Journal},
   volume = {890},
   pages = {68},
   year = {2020},
   doi = {10.3847/1538-4357/ab4f7a},
   number = {1},
-  section = {},
-  publisher = {},
   url = {https://ui.adsabs.harvard.edu/abs/2020ApJ...890...68S/abstract},
 }
-
 @article{Weberg:2023,
   doi = {10.21105/joss.04914},
   url = {https://doi.org/10.21105/joss.04914},
