+++ joss/paper.md
@@ -97,7 +97,7 @@ XRTpy's capabilities are designed to empower researchers and scientists to fully
 
 XRTpy version `0.4.0` was released on December 5, 2023. This version, available through the [Python Package Index](https://pypi.org/project/xrtpy/) (PyPI), can be installed using `pip` and is compatible with Python 3.9 and later.
 
-In fostering collaboration within the solar physics community, interoperability with other packages is a work in progress that we've started with the developers of [aiapy](https://aiapy.readthedocs.io/en/latest/) for SDO/AIA observations, [EISpack](https://eispac.readthedocs.io/en/latest/#) for Hinode-EUV imaging spectrometer (EIS) data analysis, and [irispy-lmsal](https://irispy-lmsal.readthedocs.io/en/stable/) [@Weberg:2023] for Interface Region Imaging Spectrograph (IRIS) observations. This integration provides users with a smooth and comprehensive analysis experience. Further building on the SunPy framework [@SunPy:2020], XRTpy effectively utilizes the `Map` object for handling Hinode/XRT image data.
+In fostering collaboration within the solar physics community, interoperability with other packages is a work in progress that we've started with the developers of [aiapy](https://aiapy.readthedocs.io/en/latest/) for SDO/AIA observations [@Barnes:2020], [EISpack](https://eispac.readthedocs.io/en/latest/#) for Hinode-EUV imaging spectrometer (EIS) data analysis [@Weberg:2023], and [irispy-lmsal](https://irispy-lmsal.readthedocs.io/en/stable/) for Interface Region Imaging Spectrograph (IRIS) observations. This integration provides users with a smooth and comprehensive analysis experience. Further, building on the SunPy framework [@SunPy:2020], XRTpy effectively utilizes the `Map` object for handling Hinode/XRT image data.
 
 The development of XRTpy is an open and collaborative effort, hosted on [Github](https://github.com/HinodeXRT/xrtpy) to ensure transparency and encourage community involvement. The project's documentation is comprehensive and continuously updated, available online via [Read the Docs](https://xrtpy.readthedocs.io/en/stable/). To maintain high-quality standards, XRTpy employs a robust testing framework built on [`pytest`](https://pytest.org) and [`GitHub Actions`](https://github.com/features/actions). This framework covers a range of aspects including different Python versions, online functionality, documentation integrity, software functionality, and code style checks, ensuring a reliable and effective tool for users.
 
