### 2024-07-31 e4e9d6f
Reformatted bib file
