+++ joss/paper.bib
@@ -18,13 +18,13 @@
   volume = {5},
   number = {55},
   pages = {2801},
-  author = {Will T. Barnes and Mark C. m. Cheung and Monica G. Bobra and Paul F. Boerner and Georgios Chintzoglou and Drew Leonard and Stuart J. Mumford and Nicholas Padmanabhan and Albert Y. Shih and Nina Shirman and David Stansby and Paul J. Wright},
+  author = {Barnes, Will T. and Cheung, Mark C. M. and Bobra, Monica G. and Boerner, Paul F. and Chintzoglou, Georgios and Leonard, Drew and Mumford, Stuart J. and Padmanabhan, Nicholas and Shih, Albert Y. and Shirman, Nina and Stansby, David and Wright, Paul J.},
   title = {aiapy: A Python Package for Analyzing Solar EUV Image Data from AIA},
   journal = {Journal of Open Source Software}
 }
 
 @ARTICLE{Dere:1997,
-  author = {K. P. Dere and E. Landi and H.E. Mason and B. C. Monsignori Fossi and P. R. Young},
+  author = {Dere, K. P. and Landi, E. and Mason, H. E. and Monsignori Fossi, B. C. and Young, P. R.},
   title = "{CHIANTI - An Atomic Database For Emission Lines - Paper I: Wavelengths greater than 50 Å}",
   journal = {Astronomy and Astrophysics Suppl. Ser.},
   keywords = {},
@@ -37,7 +37,7 @@
 }
 
 @ARTICLE{DelZanna:2020,
-  author = {G. Del Zanna and K.P. Dere and P.R. Young and E. Landi},
+  author = {Del Zanna, G. and Dere, K. P. and Young, P. R. and Landi, E.},
   title = "{CHIANTI - An Atomic Database for Emission Lines. Paper 16, Version 10.0, Further Extensions}",
   journal = {The Astrophysical Journal},
   keywords = {},
@@ -50,13 +50,12 @@
 }
 
 @ARTICLE{Fludra-and-Schmelz:1999,
-  author = {{Fludra}, A. and {Schmelz}, J.~T.},
-  title = "{The absolute coronal abundances of sulfur, calcium, and iron from
-Yohkoh-BCS flare spectra}",
-  journal = "{Astronomy \& Astrophysics}",
+  author = {Fludra, A. and Schmelz, J. T.},
+  title = "{The absolute coronal abundances of sulfur, calcium, and iron from Yohkoh-BCS flare spectra}",
+  journal = {Astronomy \& Astrophysics},
   keywords = {SUN: ABUNDANCES, SUN: CORONA, SUN: FLARES, SUN: X-RAYS, GAMMA RAYS},
-  year = 1999,
-  month = aug,
+  year = {1999},
+  month = {aug},
   volume = {348},
   pages = {286-294},
   adsurl = {https://ui.adsabs.harvard.edu/abs/1999A&A...348..286F},
@@ -65,7 +64,7 @@ Yohkoh-BCS flare spectra}",
 
 @article{Feldman:1992,
   title = {The Potential for Plasma Diagnostics from Stellar Extreme-Ultraviolet Observations},
-  author = {Feldman, U. and Mandelbaum, P. and Seely, J.F. and Doschek, G.A. and Gursky, H.},
+  author = {Feldman, U. and Mandelbaum, P. and Seely, J. F. and Doschek, G. A. and Gursky, H.},
   journal = {Astrophysical Journal Supplement Series},
   volume = {81},
   pages = {387},
@@ -73,18 +72,11 @@ Yohkoh-BCS flare spectra}",
   doi = {10.1086/191698}
 }
 
-
 @ARTICLE{Golub:2007,
[…]
