### 2024-07-31 14f3638
Final Updates for XRTpy JOSS Submission - Part 2 (#270)
* Updating text

* Updated summary info about XRT Obs.

* Adding Dere:1997 reference - Chianti

* Adding DelZanna:2020 reference - Chianti V10

* Adding Chianti reference

* Moved setence

* Adding log text to a plot label

* Removing Figure 3-repeated

* Adding Weberg EISpack paper

* Adding Weberg EISpack paper-forgot to add :

* Adding Weberg EISpack paper-forgot to add :

* Adding Barnes:2020 -aiapy

* Adding Sunpy correctly

* Adding Sunpy correctly

* Reformatted bib file

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for 
[…]
