+++ joss/paper.md
@@ -47,9 +47,7 @@ bibliography: paper.bib
 
 # Summary
 
-The XRTpy Python package is a specialized tool developed for the analysis of observations made by the X-Ray Telescope (XRT) [@Golub:2007] aboard the Hinode spacecraft [@Kosugi:2007].
-Hinode, a joint mission involving space agencies from Japan, the United States, and Europe. Hinode was launched with the primary aim of providing multi-wavelength data from the photosphere to the upper corona, enabling continuous observations of the Sun.
-Within this mission, the XRT instrument stands out as a remarkable piece of technology, capable of capturing high-resolution images of the solar corona's hottest material, spanning temperatures from 10^6^ to 10^7^ K.
+The XRTpy Python package is a specialized tool developed for the analysis of observations made by the X-Ray Telescope (XRT) [@Golub:2007] aboard the Hinode spacecraft [@Kosugi:2007]. Hinode is a joint mission involving space agencies from Japan, the United States, and Europe, launched with the primary aim of providing multi-wavelength data from the photosphere to the upper corona, enabling continuous observations of the Sun. Within this mission, the XRT instrument stands out as a remarkable piece of technology, capable of capturing high-resolution images of the solar corona's hottest material, diagnosing coronal temperatures from less than 1 MK to more than 10 MK.
 
 # Statement of need
 
@@ -67,7 +65,6 @@ A shift towards Python is underway within both NASA and the wider scientific com
 With XRTpy, Python users can efficiently analyze and process Hinode/XRT data, bridging the gap between traditional IDL routines and the increasing adoption of Python within the scientific community.
 This transition not only enhances accessibility but also supports the broader trend in the scientific community toward Python-based data analysis tools, thereby fostering a collaborative and efficient environment for solar researchers.
 
-
 # Package Structure
 
 XRTpy is equipped with a range of capabilities tailored for the comprehensive analysis of XRT observations.
@@ -77,7 +74,7 @@ The package is structured into distinct modules, each serving a specific purpose
 
  - `xrtpy.response.effective_area`: XRTpy's capability to calculate effective areas for various XRT filter channels, combined with CCD contamination layer thickness information, is crucial for understanding instrumental spectral responses, as depicted in \autoref{fig:Figure 1}.
 
- - `xrtpy.response.temperature_response`: XRTpy provides the capability of computing the temperature response of all the XRT filter channels. It does this by relying on a spectral emission model, drawing from [@Narukage:2011] and [@Narukage:2014]. Users can choose from a range of CHIANTI abundance sets, including the default model with coronal abundances, hybrid, and photospheric options. \autoref{fig:Figure 2} shows XRTpy's temperature response calculations for all XRT filters across the different CHIANTI  abundance sets.
+ - `xrtpy.response.temperature_response`: XRTpy provides the capability of computing the temperature response of all the XRT filter channels. It does this by relying on a spectral emission model, drawing from [@Narukage:2011] and [@Narukage:2014]. Users can choose from a range of CHIANTI abundance sets, including the default model with coronal abundances, hybrid, and photospheric options. \autoref{fig:Figure 2} shows XRTpy's temperature response calculations for all XRT filters across the different CHIANTI abundance sets. The CHIANTI database is described in [@Dere:1997] and the version used in XRTpy is CHIANTI version 10.0 [@Del_Zanna:2020]. Researchers have the flexibility to select the abundance model that best aligns with their research requirements.
 
[…]
