+++ joss/paper.bib
@@ -10,14 +10,52 @@
   note = {Volume publication date: 22 September 2009}
 }
 
+@article{Barnes:2020,
+  doi = {10.21105/joss.02801},
+  url = {https://doi.org/10.21105/joss.02801},
+  year = {2020},
+  publisher = {The Open Journal},
+  volume = {5},
+  number = {55},
+  pages = {2801},
+  author = {Barnes, Will T. and Cheung, Mark C. M. and Bobra, Monica G. and Boerner, Paul F. and Chintzoglou, Georgios and Leonard, Drew and Mumford, Stuart J. and Padmanabhan, Nicholas and Shih, Albert Y. and Shirman, Nina and Stansby, David and Wright, Paul J.},
+  title = {aiapy: A Python Package for Analyzing Solar EUV Image Data from AIA},
+  journal = {Journal of Open Source Software}
+}
+
+@ARTICLE{Dere:1997,
+  author = {Dere, K. P. and Landi, E. and Mason, H. E. and Monsignori Fossi, B. C. and Young, P. R.},
+  title = "{CHIANTI - An Atomic Database For Emission Lines - Paper I: Wavelengths greater than 50 Å}",
+  journal = {Astronomy and Astrophysics Suppl. Set.},
+  keywords = {},
+  year = {1997},
+  month = {},
+  volume = {125},
+  pages = {149-173},
+  adsurl = {https://ui.adsabs.harvard.edu/abs/1997A&AS..125..149D/abstract},
+  adsnote = {Available from ADS}
+}
+
+@ARTICLE{DelZanna:2020,
+  author = {Del Zanna, G. and Dere, K. P. and Young, P. R. and Landi, E.},
+  title = "{CHIANTI - An Atomic Database for Emission Lines. Paper 16, Version 10.0, Further Extensions}",
+  journal = {The Astrophysical Journal},
+  keywords = {},
+  year = {2021},
+  month = {},
+  volume = {909},
+  pages = {38},
+  adsurl = {https://ui.adsabs.harvard.edu/abs/2021ApJ...909...38D/abstract},
+  adsnote = {Available from ADS}
+}
+
 @ARTICLE{Fludra-and-Schmelz:1999,
-  author = {{Fludra}, A. and {Schmelz}, J.~T.},
-  title = "{The absolute coronal abundances of sulfur, calcium, and iron from
-Yohkoh-BCS flare spectra}",
-  journal = "{Astronomy \& Astrophysics}",
+  author = {Fludra, A. and Schmelz, J. T.},
+  title = "{The absolute coronal abundances of sulfur, calcium, and iron from Yohkoh-BCS flare spectra}",
+  journal = {Astronomy \& Astrophysics},
   keywords = {SUN: ABUNDANCES, SUN: CORONA, SUN: FLARES, SUN: X-RAYS, GAMMA RAYS},
-  year = 1999,
-  month = aug,
+  year = {1999},
+  month = {aug},
   volume = {348},
   pages = {286-294},
   adsurl = {https://ui.adsabs.harvard.edu/abs/1999A&A...348..286F},
@@ -26,7 +64,7 @@ Yohkoh-BCS flare spectra}",
 
 @article{Feldman:1992,
   title = {The Potential for Plasma Diagnostics from Stellar Extreme-Ultraviolet Observations},
-  author = {Feldman, U. and Mandelbaum, P. and Seely, J.F. and Doschek, G.A. and Gursky, H.},
+  author = {Feldman, U. and Mandelbaum, P. and Seely, J. F. and Doschek, G. A. and Gursky, H.},
   journal = {Astrophysical Journal Supplement Series},
[…]
