+++ joss/paper.bib
@@ -26,15 +26,17 @@
 @ARTICLE{Dere:1997,
   author = {Dere, K. P. and Landi, E. and Mason, H. E. and Monsignori Fossi, B. C. and Young, P. R.},
   title = "{CHIANTI - An Atomic Database For Emission Lines - Paper I: Wavelengths greater than 50 Å}",
-  journal = {Astronomy and Astrophysics Suppl. Set.},
-  keywords = {},
+  journal = {Astronomy and Astrophysics Suppl. Ser.},
+  keywords = {ATOMIC DATA; ASTRONOMICAL DATA BASES: MISCELLANEOUS; ULTRAVIOLET: GENERAL; SUN: ATMOSPHERE; STARS: ATMOSPHERE},
   year = {1997},
-  month = {},
+  month = {October},
   volume = {125},
   pages = {149-173},
+  doi = {10.1051/aas:1997368},
   adsurl = {https://ui.adsabs.harvard.edu/abs/1997A&AS..125..149D/abstract},
   adsnote = {Available from ADS}
 }
+
 
 @ARTICLE{DelZanna:2020,
   author = {Del Zanna, G. and Dere, K. P. and Young, P. R. and Landi, E.},
