+++ joss/paper.bib
@@ -187,14 +187,15 @@
 
 @article{SunPy:2020,
   title = {The SunPy Project: Open Source Development and Status of the Version 1.0 Core Package},
-  author = {SunPy Community and Barnes, Will T. and Bobra, Monica G. and Christe, Steven D. and Freij, Nabil and Hayes, Laura A. and Ireland, Jack and Mumford, Stuart and Perez-Suarez, David and Ryan, Daniel F. and Shih, Albert Y. and (Primary Paper Contributors) and Chanda, Prateek and Glogowski, Kolja and Hewett, Russell and Hughitt, V. Keith and Hill, Andrew and Hiware, Kaustubh and Inglis, Andrew and Kirk, Michael S. F. and Konge, Sudarshan and Mason, James Paul and Maloney, Shane Anthony and Murray, Sophie A. and Panda, Asish and Park, Jongyeob and Pereira, Tiago M. D. and Reardon, Kevin and Savage, Sabrina and Sipőcz, Brigitta M. and Stansby, David and Jain, Yash and Taylor, Garrison and Yadav, Tannmay and Rajul and Dang, Trung Kien and (Sunpy Contributors)},
+  author = {SunPy Community},
   journal = {The Astrophysical Journal},
   volume = {890},
   pages = {68},
   year = {2020},
   doi = {10.3847/1538-4357/ab4f7a},
   number = {1},
-  url = {https://ui.adsabs.harvard.edu/abs/2020ApJ...890...68S/abstract}
+  url = {https://ui.adsabs.harvard.edu/abs/2020ApJ...890...68S/abstract},
+  note = {Primary Paper Contributors: Will T. Barnes, Monica G. Bobra, Steven D. Christe, Nabil Freij, Laura A. Hayes, Jack Ireland, Stuart Mumford, David Perez-Suarez, Daniel F. Ryan, Albert Y. Shih. SunPy Contributors: Prateek Chanda, Kolja Glogowski, Russell Hewett, V. Keith Hughitt, Andrew Hill, Kaustubh Hiware, Andrew Inglis, Michael S. F. Kirk, Sudarshan Konge, James Paul Mason, Shane Anthony Maloney, Sophie A. Murray, Asish Panda, Jongyeob Park, Tiago M. D. Pereira, Kevin Reardon, Sabrina Savage, Brigitta M. Sipőcz, David Stansby, Yash Jain, Garrison Taylor, Tannmay Yadav, Rajul, Trung Kien Dang.}
 }
 
 @article{Weberg:2023,
