+++ joss/paper.bib
@@ -26,7 +26,7 @@
 @ARTICLE{Dere:1997,
   author = {Dere, K. P. and Landi, E. and Mason, H. E. and Monsignori Fossi, B. C. and Young, P. R.},
   title = "{CHIANTI - An Atomic Database For Emission Lines - Paper I: Wavelengths greater than 50 Å}",
-  journal = {Astronomy and Astrophysics Suppl. Ser.},
+  journal = {Astronomy and Astrophysics Suppl. Set.},
   keywords = {ATOMIC DATA; ASTRONOMICAL DATA BASES: MISCELLANEOUS; ULTRAVIOLET: GENERAL; SUN: ATMOSPHERE; STARS: ATMOSPHERE},
   year = {1997},
   month = {October},
