+++ joss/paper.bib
@@ -46,7 +46,7 @@
   month = {October},
   volume = {125},
   pages = {149-173},
-  doi = {10.1051/{aas}:1997368},
+  doi = {10.1051/aas:1997368},
   adsurl = {https://ui.adsabs.harvard.edu/abs/1997A&AS..125..149D/abstract},
   adsnote = {Available from ADS}
 }
