+++ joss/paper.bib
@@ -51,8 +51,6 @@
   adsnote = {Available from ADS}
 }
 
-
-
 @ARTICLE{DelZanna:2020,
   author = {Del Zanna, G. and Dere, K. P. and Young, P. R. and Landi, E.},
   title = "{CHIANTI - An Atomic Database for Emission Lines. Paper 16, Version 10.0, Further Extensions}",
