### 2024-07-31 3f96d52
Final Updates for XRTpy JOSS Submission - Part 3 (#271)
* Correcting @DelZanna:2020 reference

* Adding DOI to Dere

* Fixing Sunpy reference

* Updating note

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Fixing aas code spell

* Updating ignore-wors-list-adding aas

* Updating aas

* Fix spacing

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
