+++ joss/paper.bib
@@ -27,11 +27,26 @@
   author = {Dere, K. P. and Landi, E. and Mason, H. E. and Monsignori Fossi, B. C. and Young, P. R.},
   title = "{CHIANTI - An Atomic Database For Emission Lines - Paper I: Wavelengths greater than 50 Å}",
   journal = {Astronomy and Astrophysics Suppl. Set.},
-  keywords = {},
+  keywords = {ATOMIC DATA; ASTRONOMICAL DATA BASES: MISCELLANEOUS; ULTRAVIOLET: GENERAL; SUN: ATMOSPHERE; STARS: ATMOSPHERE},
   year = {1997},
-  month = {},
+  month = {October},
   volume = {125},
   pages = {149-173},
+  doi = {10.1051/aas:1997368},
+  adsurl = {https://ui.adsabs.harvard.edu/abs/1997A&AS..125..149D/abstract},
+  adsnote = {Available from ADS}
+}
+
+@ARTICLE{Dere:1997,
+  author = {Dere, K. P. and Landi, E. and Mason, H. E. and Monsignori Fossi, B. C. and Young, P. R.},
+  title = "{CHIANTI - An Atomic Database For Emission Lines - Paper I: Wavelengths greater than 50 Å}",
+  journal = {Astronomy and Astrophysics Supplement Series},
+  keywords = {ATOMIC DATA; ASTRONOMICAL DATA BASES: MISCELLANEOUS; ULTRAVIOLET: GENERAL; SUN: ATMOSPHERE; STARS: ATMOSPHERE},
+  year = {1997},
+  month = {October},
+  volume = {125},
+  pages = {149-173},
+  doi = {10.1051/aas:1997368},
   adsurl = {https://ui.adsabs.harvard.edu/abs/1997A&AS..125..149D/abstract},
   adsnote = {Available from ADS}
 }
@@ -185,14 +200,15 @@
 
 @article{SunPy:2020,
   title = {The SunPy Project: Open Source Development and Status of the Version 1.0 Core Package},
-  author = {SunPy Community and Barnes, Will T. and Bobra, Monica G. and Christe, Steven D. and Freij, Nabil and Hayes, Laura A. and Ireland, Jack and Mumford, Stuart and Perez-Suarez, David and Ryan, Daniel F. and Shih, Albert Y. and (Primary Paper Contributors) and Chanda, Prateek and Glogowski, Kolja and Hewett, Russell and Hughitt, V. Keith and Hill, Andrew and Hiware, Kaustubh and Inglis, Andrew and Kirk, Michael S. F. and Konge, Sudarshan and Mason, James Paul and Maloney, Shane Anthony and Murray, Sophie A. and Panda, Asish and Park, Jongyeob and Pereira, Tiago M. D. and Reardon, Kevin and Savage, Sabrina and Sipőcz, Brigitta M. and Stansby, David and Jain, Yash and Taylor, Garrison and Yadav, Tannmay and Rajul and Dang, Trung Kien and (Sunpy Contributors)},
+  author = {SunPy Community},
   journal = {The Astrophysical Journal},
   volume = {890},
   pages = {68},
   year = {2020},
   doi = {10.3847/1538-4357/ab4f7a},
   number = {1},
-  url = {https://ui.adsabs.harvard.edu/abs/2020ApJ...890...68S/abstract}
+  url = {https://ui.adsabs.harvard.edu/abs/2020ApJ...890...68S/abstract},
[…]
