+++ joss/paper.bib
@@ -200,7 +200,7 @@
 
 @article{SunPy:2020,
   title = {The SunPy Project: Open Source Development and Status of the Version 1.0 Core Package},
-  author = {SunPy Community},
+  author = {SunPy-Community},
   journal = {The Astrophysical Journal},
   volume = {890},
   pages = {68},
