### 2024-08-01 #6396 @joyvelasquez
Hello @dfm ,

I'm also most done and soon to publish on Zendo, working on metadata at the moment. However, I was able to generate/reserve a DOI tag from Zendo.

The version number of the software has been incremented to 0.4.1.
The DOI for the archived release is 10.5281/zenodo.13157914.
