### 2024-08-04 #6396 @dfm
Thanks @joyvelasquez! It looks like that DOI doesn't resolve. Please share the DOI once you have one with the metadata set up and then the final processing will be quick. Thanks!
