### 2024-08-05 34d4585
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/python-jsonschema/check-jsonschema: 0.29.0 → 0.29.1](https://github.com/python-jsonschema/check-jsonschema/compare/0.29.0...0.29.1)
- [github.com/astral-sh/ruff-pre-commit: v0.5.4 → v0.5.6](https://github.com/astral-sh/ruff-pre-commit/compare/v0.5.4...v0.5.6)
- [github.com/nbQA-dev/nbQA: 1.8.5 → 1.8.7](https://github.com/nbQA-dev/nbQA/compare/1.8.5...1.8.7)
