### 2024-08-08 #6396 @joyvelasquez
Hello @dfm , I managed to finish up metadata on Zenodo, and published XRTpy.
The DOI for the archived release is  : https://zenodo.org/records/13157914
