### 2024-08-08 #6396 @joyvelasquez
Hello Dan, Stuart, & David,

I want to express my deepest gratitude to all of you for your invaluable contributions to the review and publication of XRTpy in JOSS. Your thoughtful feedback, constructive suggestions, and attention to detail have significantly improved the quality of the package and its documentation. @dfm, your guidance as the editor was instrumental in navigating the submission process smoothly. @Cadair and @dpshelio, your insightful reviews have provided new perspectives and helped shape XRTpy into a more robust and user-friendly tool.

I am truly thankful as I speak on behalf of the XRTpy team, for the time and effort you have each dedicated to this project, and I’m excited to see how XRTpy will benefit the solar physics community.
