### 2024-08-09 #6396 @dfm
@joyvelasquez — It looks like the `Dere:1997` reference is defined twice in the bibliography. Can you remove one of those? Sorry I didn't catch this sooner!
