### 2024-08-09 4f6681b
Update paper.bib
Dere:1997 reference is defined twice in the bibliography, a reference was removed.
