### 2024-08-09 #6396 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
