### 2024-08-09 #6396 @dfm
@dpshelio, @Cadair — Many thanks for your reviews here! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you!!

Note: It looks like the PDF isn't rendering (at least for me) at the DOI. This happens sometimes because of caching issues with our servers. Let's check back tomorrow if it doesn't resolve automatically.
