### 2024-08-10 #6396 @Cadair
Thanks @dfm always happy to review JOSS papers :)
