### 2024-08-10 #6396 @Cadair
@joyvelasquez Thanks for writing XRTPy! Don't forget to add XRTPy to the list of sunpy affiliated packages :)
